### 2022-03-26 #166 @whedon
[PRE REVIEW]: Spatial data analysis with R: wrangling, visualization and econometric models
**Submitting author:** @jaime-pru (<a href="http://orcid.org/0000-0001-5866-1280">Jaime Alberto Prudencio Vázquez</a>)
**Repository:** <a href="https://github.com/jaime-pru/Analisis-de-datos-espaciales" target ="_blank">https://github.com/jaime-pru/Analisis-de-datos-espaciales</a>
**Version:** 1.0

Due to the challenges of the COVID-19 pandemic, JOSE is currently operating in a "reduced service mode".
