### 2022-04-19 #173 @whedon
[REVIEW]: Spatial data analysis with R: wrangling, visualization and econometric models
**Submitting author:** @jaime-pru (<a href="http://orcid.org/0000-0001-5866-1280">Jaime Alberto Prudencio Vázquez</a>)
**Repository:** https://github.com/jaime-pru/Analisis-de-datos-espaciales
**Branch with paper.md** (empty if default branch):
**Version:** 1.0

- [ ] **Version:** Does the release version given match the repository release (1.0)?

- [ ] **Version:** Does the release version given match the repository release (1.0)?
