### 2022-05-03 #173 @nmorandeira
Hi @whedon I’m revising the submission (I’ve already revised the paper and
now I’m going through the last two chapters of the course).
Can you remind me the deadline?

El El mar, 3 de may. de 2022 a la(s) 18:32, whedon ***@***.***>
escribió:

> 👋 @nmorandeira <https://github.com/nmorandeira> , please update us on
> how your review is going (this is an automated reminder).
>
