### 2022-05-11 jaime-pru/Analisis-de-datos-espaciales#1 @nmorandeira
Community guidelines: contributions to the repository // Guías a la comunidad: contribuciones al repositorio [in English with suggestion in Spanish] // JOSE Review
[_ES] Este _issue_ se vincula a la revisión del artículo para la revista [JOSE Review](https://github.com/openjournals/jose-reviews/issues/173#event-6459353208)
[EN] This issue is linked to the [JOSE Review](https://github.com/openjournals/jose-reviews/issues/173#event-6459353208)_

These are suggestions and comments to improve community guidelines.

How do you encourage students to contribute to the repository? Although you state _"On the other hand, the student or teacher reader interested in making some correction, modification or contribution to this course, can do so through the instructions located in the repository that contains the source files of the electronic book_":

- There are no issues or PR yet. Was March-September 2020 the only edition of the course?

- It is difficult to find the instructions located in the repository. I suggest linking them to the README.md file.
Are these ones? https://github.com/jaime-pru/Analisis-de-datos-espaciales/blob/main/Contributing.md

- The instructions are in English, while your students are Spanish speakers and (as you pointed out) may have no English knowledge. I would add a Spanish text on how to contribute, which may be a translation of the link on Data School. However, creating a pull request is difficult for new github users. I would suggest two or three options to contribute, according to the student's github level, something like this:

> a) Envíame un correo a esta dirección: xxxx.
>
> b) Deja un comentario en un "Issue" de gitbub. Necesitas crear una cuenta en github, ir [aquí](https://github.com/jaime-pru/Analisis-de-datos-espaciales/issues) y hacer click en "New issue". Luego simplemente escoges un título, completas tu comentario y lo envías con el botón "Submit new issue".
>
> c) Sugerir los cambios a través de un Pull Request en el repositorio. Necesitas una cuenta de github y seguir [estas instrucciones](https://www.dataschool.io/how-to-contribute-on-github/) (en inglés). Lleva un poco más de trabajo y de conocimiento de github que la opción (b).
>
> Cualquier contribución será bienvenida.

- Minor: "let me know".
