### 2022-05-11 jaime-pru/Analisis-de-datos-espaciales#2 @nmorandeira
Accesibilidad e inclusión / Accessibility and inclusion [in Spanish] // JOSE Review
_[ES] Este _issue_ se vincula a la revisión del artículo para la revista [JOSE Review](https://github.com/openjournals/jose-reviews/issues/173#event-6459353208)
[EN] This issue is linked to the [JOSE Review](https://github.com/openjournals/jose-reviews/issues/173#event-6459353208)_

Las siguientes son recomendaciones para mejorar la accesibilidad del curso.

- Las imágenes deberían tener texto alternativo, para que puedan ser accesibles para personas ciegas o con visión reducida, que usan lectores de pantalla. También es más accesible para personas con mala conectividad de internet. En caso de no poder colocar alt-tex en bookdown (creo que sí se puede), sugiero agregarlo en el epígrafe de cada imagen luego del título de la figura.

- En caso de usar una tabla o cuadro, incluirla con formato html y no como imagen. Ejemplo Figura 1.2.

- Agregar el título del curso en texto (que no esté sólo en la imagen de portada en el Prólogo). Además de mejorar la accesibilidad del curso, permite que los buscadores indexen el curso. Si bien hay un título arriba a la izquierda sobre la barra de búsqueda, el título es incompleto en algunas de las páginas (falta la bajada: Manejo, visualización y modelos econométricos).

- En lo posible, traducir los subtítulos: References, Table of contents, On this page. Search en la barra de búsqueda. El footer también está en inglés “was written by”, etc. Copy to clipboard, en los códigos.

- Cambiar el link de CC BY 4.0 al pie de la página por su equivalente en castellano: https://creativecommons.org/licenses/by/4.0/deed.es

- En el capítulo 1, hay un ejemplo que presenta categorías binarias de sexo, “femenino” y “masculino”. El ejemplo puede ser excluyente de personas intersex, no binarias o que no se identifican con estas categorías. Dado que es un ejemplo irrelevante para el curso, lo cambiaría por otro que también permita explicar las variables categóricas. Seguramente hay un ejemplo relacionado con covid19 que sea útil (caso positivo/negativo, vacunado/no vacunado, etc).

- Recursos interesantes sobre la redacción de alt-text y accesibilidad: esta charla https://www.youtube.com/watch?v=DxLkv2iRdf8 y su repositorio con abundantes recursos vinculados https://github.com/spcanelon/csvConf2021
