### 2022-05-11 jaime-pru/Analisis-de-datos-espaciales#3 @nmorandeira
Prólogo / Prologue [in Spanish] // JOSE Review
_[ES] Este _issue_ se vincula a la revisión del artículo para la revista [JOSE Review](https://github.com/openjournals/jose-reviews/issues/173#event-6459353208)
[EN] This issue is linked to the [JOSE Review](https://github.com/openjournals/jose-reviews/issues/173#event-6459353208)_

Listo aquí unas sugerencias que creo que podrían mejorar el prólogo.

- Agregar el título del curso en texto (que no esté sólo en la imagen de portada).

- “Ciencias regionales”. Incluir definición y/o una referencia. En mi caso soy bióloga y trabajo con datos espaciales a escala regional y de paisajes; no conozco exactamente a qué se denomina “ciencias regionales”. La cita de Fisher y Nijkamp 2014, si bien podría ser una definición, se presenta como “han ido adquiriendo una amplia orientación…” y entonces parece hablarnos de un cambio en las características de las ciencias regionales que aún no definimos.

- En el segundo párrafo, sugiero mencionar que el abordaje de los tópicos de las ciencias regionales requieren trabajar con datos o información espacial; que la localización geográfica de los procesos debe ser explícita. Eso da sentido al curso. Si no, sólo tácitamente conocemos que las ciencias regionales se vinculan con los datos espaciales, porque suponemos que el título del curso está relacionado con el prólogo.

- Tercer párrafo, _Handbook of Regional Science_: incluir año; sugiero agregar un equivalente en castellano entre paréntesis.

- Tercer párrafo: Aclarar que la universidad queda en México.

- No comprendo por qué el enlace de la portada lleva a un curso de actualización que parece ser otro distinto. Quitaría el enlace. En caso contrario, lo incluiría al final del prólogo aclarando de qué se trata y por qué sería interesante visitar ese sitio.

- De algunos autores mencionados se incluye el nombre propio, de otros sólo la inicial. Uniformizar.

- Quinto párrafo. Agregaría que R es un software libre, ya que luego se enfatiza este punto al final del prólogo. Puede ser una mención muy breve, y quizás mencionar que se abordará en el Capítulo 1.

- Sexto párrafo. Mencionar en mayúsculas los capítulos y enlazar a sus páginas.

- Al final del prólogo, o luego de mencionar la licencia, se puede incluir link a la explicación de cómo contribuir (ver issue  https://github.com/jaime-pru/Analisis-de-datos-espaciales/issues/1).

- Si bien considero que es decisión del autor usar lenguaje no sexista o inclusivo, creo que en otras partes del libro se ha enfatizado la mención de mujeres (ejemplo: autor o autora del paquete readxl). Si ese es el espíritu tuyo como autor, en el prólogo sería atinado usar también el género femenino, ya que es la primera aproximación al libro.
