### 2022-05-11 jaime-pru/Analisis-de-datos-espaciales#4 @nmorandeira
Capítulo 1. R: Una introducción (...) / Chapter 1 [in Spanish]
_[ES] Este _issue_ se vincula a la revisión del artículo para la revista [JOSE Review](https://github.com/openjournals/jose-reviews/issues/173#event-6459353208)
[EN] This issue is linked to the [JOSE Review](https://github.com/openjournals/jose-reviews/issues/173#event-6459353208)_

Las siguientes son recomendaciones para el Capítulo 1.

- Se explica muy detalladamente cómo descargar R y RStudio. Aclarar que primero hay que instalar R y luego RStudio.

- Uniformizar: RStudio y no R Studio

- Figura 1.1; Figura 1.2, etc. En mayúscula.

- Sugiero cambiar la Figura 1.1. por el modo claro, ya que el modo oscuro no es el que se ve por defecto en una instalación fresca.

- La Figura 1.2 en verdad es una Tabla/Cuadro 1.1. Sugiero cambiarlo a formato tabla ya que es más claro (como lectora esperaba una figura y no una tabla) y además mejora la accesibilidad.

- Cambiar “Hello world” por “Hola mundo” u otro texto (“Hola México”?).

- “O simplemente da click en el ícono de copiar…”. Faltaría indicar que hay que pegar el texto copiado en la consola.

- No se mencionó que # significa comentario y no se corre en R

- Obtener ayuda de un paquete: se usa un único signo de interrogación ( help(tmap) es equivalente a ?tmap, no a ??tmap). Se usan dos signos para una función dentro de un paquete.

- Se menciona el libro R for Data Science. Sugiero cambiar por la versión [**R para Ciencias de Datos**](https://es.r4ds.hadley.nz/). Fue traducido a castellano por la comunidad de #RLadies de Latinoamérica y es excelente.

- Se mencionan también las Cheatsheets, lo cual me parece excelente porque tardé mucho en descubrirlas! Agregaría que aquí hay versiones de muchas de ellas traducidas al castellano, incluso está la de ggplot2 que se menciona en el curso https://www.rstudio.com/resources/cheatsheets/#translations

- Si usás la notación readxl::read_excel, no es necesario usar library(readxl).

- Agregaría la explicación de cómo citar un paquete (citation (“tmap”). Quizás cuando se introducen los paquetes, se puede decir que es importante citarlos si los usamos.

**Errores ortotipográficos:**

- “RStudio puede ser **descargado** de”

- “**si** colocas el número 5” (sin tilde).

- “**Estos** últimos” (sin tilde)

- “se utiliza el **signo**”

- “trabajo: la” (sin mayúscula después de dos puntos)

- “femenino”

- “como el que se mencionó antes” (minúscula en “como” y sobra un “en”).
