### 2022-05-11 jaime-pru/Analisis-de-datos-espaciales#5 @nmorandeira
Capítulo 2. Mapas coropléticos (...) / Chapter 2  [in Spanish] // JOSE Review
_[ES] Este _issue_ se vincula a la revisión del artículo para la revista [JOSE Review](https://github.com/openjournals/jose-reviews/issues/173#event-6459353208)
[EN] This issue is linked to the [JOSE Review](https://github.com/openjournals/jose-reviews/issues/173#event-6459353208)_

Las siguientes son recomendaciones para el Capítulo 2:

- **¿Cuentas  con permiso para reproducir el texto de Coursera? “What is data”.** Por favor chequear la licencia del curso. En caso de estar permitida la reproducción, citar la fuente en el bookdown (no sólo en el pdf). Recordar también que un html es más accesible que un pdf para personas que usan lectores de pantalla.

- Considero que una referencia que no puede faltar en un curso de datos espaciales en castellano es el libro de Víctor Olaya. Creo que es el mejor material teórico sobre SIG. Encima está escrito en castellano por un español y es un libro libre :) https://volaya.github.io/libro-sig/

- Se introduce el término “georreferenciada” sin explicarlo ni definirlo. Resulta un concepto sumamente importante para el curso.

- **Deben revisarse las definiciones de datos vectoriales y ráster.**

En cuanto a _datos vectoriales_, la actual redacción parece decir que se usan puntos, líneas y polígonos en un mismo objeto espacial, cuando en verdad estos tipos de representaciones no pueden mezclarse en un mismo vectorial.

En cuanto a los _datos ráster,_ la definición es incorrecta: no necesariamente los ráster provienen de imágenes tomadas por instrumentos ópticos; también pueden ser originadas por otros instrumentos (instrumentos satelitales de microondas activas o pasivas, por ejemplo); pero más generalmente un ráster puede ser una grilla que no derive de ningún instrumento sino que sea generada por un usuario, ya sea definiendo celdas y sus valores manualmente o a través de un modelo; también se pueden derivar rásters de otros rásters por álgebra de mapas. En el libro de Olaya se definen muy bien los modelos de representación vectoriales y ráster https://volaya.github.io/libro-sig/chapters/Tipos_datos.html

- Se menciona que los vectoriales de tipo SHP fueron desarrollados por ESRI. Si bien es correcto que esta empresa creó el formato, la frase es confusa ya que puede darse a entender que todos los archivos SHP son desarrollados por ESRI, incluso los que se usan en el curso.

- Explicaría la importancia del archivo .prj. Además, al descargar el shapefile del ejemplo hay un .prj. ¿Cómo se espera que un estudiante responda a la pregunta _Para qué sirven esos archivos_?

- _sf_: la explicación del paquete no es muy pertinente para un público novato que no conoce qué es GDAL, GEOS y PROJ. Mejor diría que es un paquete útil para manejar datos espaciales de tipo vectorial. En general, me parece mejor que uses tus propios términos para hablar del paquete, en lugar de reproducir lo que dice la ayuda (mismo ejemplo para tmap).

- Sección 2.4. Definir “mapa coroplético” (está en el título del capítulo y de la sección; y luego se habla de “mapa temático”, que tampoco se define).

- Sección 2.5. Paleta de colores. Sugiero incluir una mención sobre cuándo usar una escala divergente y cuándo no; también sugiero incluir una mención a la accesibilidad en cuanto al uso de colores apropiados para personas daltónicas o con ceguera de color. El capítulo 5 del libro de Víctor Olaya sobre visualización es muy interesante para sumar como bibliografía de esta sección. Por ejemplo, en esta sección versa sobre escalas divergentes y no divergentes: https://volaya.github.io/libro-sig/chapters/Mapas.html

- Sección 2.6. Para que la leyenda de los mapas quede en castellano, sugiero indicar cómo cambiar los nombres de las categorías (en particular para que diga “0 a 5” en vez de “0 to 5”).

- Nuevamente es decisión del autor, pero sugiero agregar también “entusiastas programadoras”. Justo en ese ejemplo parece que sólo los hombres desarrollan para R.

**Errores ortotipográficos:**
- “básicas de los datos y los datos espaciales” (sobra un “los”).
- “Instagram” (en un link).
- “[Maps in R](https://www.jstatsoft.org/article/view/v084i06) ([Tennekes 2018](https://jaime-pru.github.io/Analisis-de-datos-espaciales/references.html#ref-Tennekes2018))” (falta un espacio antes de la cita)
- “uno para la clase final, es decir a manera de anclas.” (se lee raro el “es decir” al final).
- “tú mismo profundices” (falta tilde en tú)
