### 2022-05-11 jaime-pru/Analisis-de-datos-espaciales#6 @nmorandeira
Capítulo 3. Análisis espacial I: autocorrelación / Chapter 3  [in Spanish] // JOSE Review
_[ES] Este _issue_ se vincula a la revisión del artículo para la revista [JOSE Review](https://github.com/openjournals/jose-reviews/issues/173#event-6459353208)
[EN] This issue is linked to the [JOSE Review](https://github.com/openjournals/jose-reviews/issues/173#event-6459353208)_

Las siguientes son recomendaciones para el Capítulo 3

- Sugiero reordenar la introducción del capítulo. Se comienza hablando de autocorrelación espacial, sin embargo lo primero que se enfatizan son ejemplos de análisis espacial (no necesariamente autocorrelación). Por ejemplo, el caso de Jon Snow no es necesariamente un ejemplo de autocorrelación espacial, ya que el principal descubrimiento fue que había dos variables asociadas (la distribución de las bombas de agua y la distribución de los casos de cólera). Es decir, sugiero organizar la introducción para hablar primero de casos en que es útil analizar el patrón espacial de una o más variables, y luego continuar con el caso específico de la autocorrelación espacial.

- En la actual forma de presentación del capítulo, se menciona a la autocorrelación espacial en la primera oración, pero sin embargo recién se la define en el cuarto párrafo. Si el autor decide no reordenar la introducción como sugerí en el párrafo anterior, sería conveniente anticipar la definición de autocorrelación espacial al primer párrafo. La definición de Chasco es interesante y adecuada. Quizás para un estudiante novato/a no queda claro por qué es “auto” correlación, por lo que sugiero enfatizar que se trata de la relación entre los valores de un indicador en un sitio y del mismo indicador en otro sitio cercano.

- Sección 3.2. Al igual que lo indicado para el capítulo 2 en otro issue, no resulta muy útil la definición de rgdal extraída de la documentación.

- Sección 3.4. Creo que por error se le dio formato de sección a un texto normal de la sección 3.3. Revisar.

- Acrónimo de LISA: usar mayúsculas; incluir traducción a castellano.

- Última figura de Sección 3.5: no se ven polígonos de las categorías azules (si no existen, aclararlo en el texto). Además agregaría un título a la leyenda, para mayor claridad.

**Errores orto-tipográficos**

- “De ser así, ¿cómo lo harías?” (sobra un “de ser así”)

- “Alto-Bajo” en leyenda de figura (dice “Alto-Bbajo”). Además “Bajo-Alto” está en minúscula”
