### 2022-05-11 jaime-pru/Analisis-de-datos-espaciales#7 @nmorandeira
Capítulo 4. Modelos de regresión lineal / Chapter 4  [in Spanish] // JOSE Review
_[ES] Este _issue_ se vincula a la revisión del artículo para la revista [JOSE Review](https://github.com/openjournals/jose-reviews/issues/173#event-6459353208)
[EN] This issue is linked to the [JOSE Review](https://github.com/openjournals/jose-reviews/issues/173#event-6459353208)_

Las siguientes son recomendaciones para el Capítulo 4

- **¿Cuentas  con permiso para reproducir el texto de Coursera? “Types of data science questions”**. Por favor chequear la licencia del curso. En caso de estar permitida la reproducción, citar correctamente la fuente en el bookdown (no sólo en el pdf): el autor de la cita en el .bib es incorrecto (aparece University 2020 como autor; cuando esa es la institución, ¿quién es el/la autor/a? Parece ser Jeffrey Leek, pero chequear).  Recordar también que un html es más accesible que un pdf para personas que usan lectores de pantalla.

- Al principio de la sección 4.5., sugiero enfatizar que este capítulo no aborda un análisis espacial, ya que no es tenida en cuenta la localización geográfica de los puntos con distintos valores del indicador.

- ¿Es correcta la cita: [C., A., y Geoffrey 2012](https://jaime-pru.github.io/Analisis-de-datos-espaciales/references.html#ref-Montgomery2012)? (parecen estar mal los autores).

- Cita de Dalpiaz. Incluir fecha. Puede ser indicada la fecha de consulta de la publicación en línea.

- Sección 4.2. Definir modelo econométrico (¿qué diferencias hay con modelos no econométricos?). Especificar en qué se diferencia un modelo econométrico espacial de una regresión lineal común.

- Sección 4.5. Me parece importante al final de esta sección mostrar el plot de poshab vs ss, con la línea de ajuste del modelo lm. Hay un ejercicio que plantea esto, quizás se puede mover más arriba a esta sección para mayor claridad del modelo obtenido.

**Errores orto-tipográficos:**

- deberían ser modestas” (falta comilla de cierre)

- En algunas partes del libro se usa “sólo” con tilde y en otras “solo” sin tilde. Uniformizar.

- “ajustados, ¿puedes” (falta coma)
