### 2022-05-11 jaime-pru/Analisis-de-datos-espaciales#8 @nmorandeira
Capítulo 5. Análisis espacial II / Chapter 5  [in Spanish] // JOSE Review
_[ES] Este _issue_ se vincula a la revisión del artículo para la revista [JOSE Review](https://github.com/openjournals/jose-reviews/issues/173#event-6459353208)
[EN] This issue is linked to the [JOSE Review](https://github.com/openjournals/jose-reviews/issues/173#event-6459353208)_

Los siguientes son comentarios sobre el Capítulo 5.

- Este capítulo tiene una carga cognitiva más alta que los demás. Se avanza rápido en conceptos muy complejos, en particular en la introducción teórica. Sugiero narrar un poco más qué utilidad tiene cada modelo, explicar con más detenimiento las opciones de la Figura 5.1, explicar un poco las ecuaciones incluidas, etc.

- No queda claro qué es el multiplicador de Lagrange (no se define ni se explica con detenimiento); por lo tanto tampoco queda claro qué son las versiones robustas de ese multiplicador.

- No queda claro por qué el modelo es “anidado”.

- Figura 5.1. Es ilegible. Texto borroso en baja resolución.

- Figuras 5.2 y 5.3. Son difíciles de leer, texto borroso y con sombreado.

- Cita Anselin y Rey. En algunos casos, no está linkeada a su referencia. Corregir apariciones de Ansuelin por Anselin.

- “hay que mirar los resultados de las versiones robustas para tomar la decisión final”, ¿por qué?

- Para mayor claridad, ¿es posible incluir gráficas que ayuden a interpretar el output de los modelos? (quizás no, pero noto contraste respecto a la interpretación de los capítulos anteriores). Se siente que el capítulo termina apresuradamente, quizás con más interpretación sobre el/los modelos seleccionados (ayudado o no con una gráfica) es posible comprender mejor los últimos resultados y quedarnos con una imagen más acabada del curso.

- No me queda claro si los modelos enseñados son específicamente econométricos. Parece ser que no, que son modelos espaciales aplicados a econometría, pero que podrían ser aplicados a otras disciplinas. Aclarar.

- Al final de este capítulo o en uno nuevo, sugiero incluir una sección con conclusiones, hilando los aportes de todo el curso, y otras aproximaciones que pueden tener lxs estudiantes si quisieran produndizar en sus análisis. Si bien esto está dicho en este capítulo, se siente que se acelera el fin del curso y sería bueno recapitular para el cierre con los aportes de cada capítulo.

**Errores orto-tipográficos**

- “es sintetizado por **nuestro autores**” (corregir, no sé a qué autores se refiere antes de ver el epígrafe).

- responder a la pregunta: “¿Los errores de nuestro modelo están correlacionados?” (agregar comillas o cursiva)
