### 2022-05-11 #173 @nmorandeira
This is my review for the  JOSE submision #173 by Jaime Alberto Prudencio Vázquez (@jaime-pru), entitled **"Spatial data analysis with R: wrangling, visualization and econometric models".** I was invited to review the submission by the editor @allisonhorst.

I've revised the article in paper.md and the bookdown course.
The course is very interesting, and the paper emphasizes its need –in particular for Spanish-speaking students (and I agree with the author in his statement of need). The course covers several topics on spatial data analysis, focused on covid19 data for Mexico and socio-demographic/ economic variables. I consider that this course fits the scope of the JOSE and would be of interest to the journal's readers. Here I provide several general comments, complemented by eight issues in the course repository (linked below).

----

**1) Comments are about the repository**, following the Review checklist:
- The course is published in a public repository.
- **Version:** the repository has no release number; the author should generate a version number.
- **License:** the repository has a license. I suggest including a link to it in the README.md
- **Other suggestions for the repository's README.md**: add a Spanish version and the course title (in Spanish and English); if possible, add a Spanish summary. Add a link to the course (there is a circular link to the repo).
- **Community guidelines:** I suggest improving these guidelines. See this issue in the course repo: https://github.com/jaime-pru/Analisis-de-datos-espaciales/issues/1

-----

**2) Comments and suggestions for the article in "paper.md "**

- In "Story of the project", tell the readers how many editions of the course you have taught; and how evolved your course through the promotions. Do you plan to grow the course into more chapters?

- **Course schedule**. How long would you expect a student to complete each chapter or the entire course? I suggest adding this information to the paper.md, but also adding an expected time in the bookdown course. For example, see the schedule in [this course]("https://datacarpentry.org/r-intro-geospatial/”) as a reference. I also suggest adding breaks, to reduce the cognitive load and allow students to learn new chunks of information.

- The book includes several **Exercises**. Can the student/reader find an answer to these exercises in the book? (do you plan to add the Solutions?). Or is the course expected to be completed with a teacher assistant? Please clarify this.

- I suggest adding some information on how someone would adopt the course, and include examples of how to use it.

- I suggest adding the Spanish title of the course, maybe at the end of the first sentence (in parentheses).

- Should the link in the first sentence point to the course webpage instead of the repo?

- Plural: "skills for the quantitative **analyses**"

- Capitalize _Chapter_ when it's followed by the chapter number. E.g. "In Chapter 1,"; "Meanwhile, in Chapter 4"; "Finally, in chapter 5,"; "from Chapter 1 to 5"

- Uniformize how you name your target: the "student" or the "reader" or the "user". I'd suggest "the student".

- Replace "database about information on the situation" with "database on the situation".

- Uniformize "México" or "Mexico".

- "In addition, the **used database** provides".

- Capitalize: "Bachelor's degree"; "Bachelor level".

- Uniformize the name of the University. I suggest using the first mention in Spanish followed by the translation between parentheses and all subsequent mentions in Spanish. Otherwise, vice-versa: first mention in English (Spanish) and continuing in English.

- The word "related" is followed by "to"; and "associated" is followed by "with". Thus, I suggest rephrasing: "Each chapter contains both theoretical and practical elements **related to dealing with** spatial data in the context of economics."

- Non-sexist language. I like your option to use "she can continue", and as far as JOSE does not state another style, it's your decision. Some English speakers use "they" as the "inclusive" pronoun (including she/he/non-binary identities); for example: "if the student feel comfortable (...), they can continue (...)". Once again, it's your decision :) and "she can continue" it's fine for me.

- "**facilitates** the learning process."

- "All **the contributions** will be welcome."

- Add a space: "Rodríguez ("

-----

**3) Comments and suggestions about the course**

I've reviewed the course and have filed one issue per chapter, with general comments and suggestions, in Spanish.

I'd like to highlight these points:

- Add a **schedule** or estimated time to complete each chapter / subchapter (see previous comment).

- **Community guidelines:** I suggest improvements in this section. See
Issue 1 (in English with suggestions in Spanish). https://github.com/jaime-pru/Analisis-de-datos-espaciales/issues/1

- **Accessibility and inclusion.** Some issues should be addressed to improve accessibility and inclusion. See Issue 2 (in Spanish). https://github.com/jaime-pru/Analisis-de-datos-espaciales/issues/2

- The prologue and Chapters 1 to 4 are easy to follow, probably guided by a teacher assistant or giving the students the possibility to ask questions in an online forum.

A general comment is to pay attention to the first mentions of the terms and add a clear **definition of the new concepts** before continuing with comparisons and further information about the concept (specific examples of this are provided in the repo issues). Several definitions are lacking. **The definitions for vector and raster data (Chapter 2) need to be revised** (see the specific issue). A glossary would be helpful, but it's not mandatory (maybe an improvement for future versions of the course).

Regarding the **references**, these are my main comments. Due to the target audience of the course, I suggest including links to the Spanish versions of the R4DS book and the RStudio Cheatsheets, instead of the English links; I also strongly suggest including Victor Olaya's open-source GIS book (in Spanish) as a reference. Please double-check that you can include Coursera's texts; if so, check how to cite their materials.

**The issues related to these chapters are:**

Links issues 3 to 7 (in Spanish):

- - Prologue https://github.com/jaime-pru/Analisis-de-datos-espaciales/issues/3
- - Chapter 1 https://github.com/jaime-pru/Analisis-de-datos-espaciales/issues/4
- - Chapter 2 https://github.com/jaime-pru/Analisis-de-datos-espaciales/issues/5
- - Chapter 3 https://github.com/jaime-pru/Analisis-de-datos-espaciales/issues/6
- - Chapter 4 https://github.com/jaime-pru/Analisis-de-datos-espaciales/issues/7

- **Chapter 5** has a high cognitive load and, in my opinion, needs major improvements. I consider that this chapter has fewer details on the explanations and examples than the previous chapters. Figure 5.1. cannot be read; Figures 5.2 and 5.3 are difficult to read. Comments on this chaper are provided in its specific issue, specifically on terms that are not defined or are explained with few details for novate students.

Links issue 8 (in Spanish):
- - Chapter 5 https://github.com/jaime-pru/Analisis-de-datos-espaciales/issues/8

- I suggest adding a **_Conclusions_ section**, recapitulating the main learnings of each chapter and some guide for curious students that wish to continue learning. Although some of this is included in the last paragraph of Chapter 5 (in Section 5.3.2, Spatial error model), I think the course deserves a nicer ending, highlighting all we have learned through the course. Also, students can find a list of references with additional material (most of them were included in Chapters 1-5, but a highlight and commented reference list could be useful). And, why not?, students and contributors can be encouraged to contribute with more spatial analyses.

----

I'm available to the author and the editor for any query about my review. Thanks for the invitation!: I have learned about pedagogy and spatial modeling while conducting this review. I've also found interesting the JOSE reviewing mechanism, as well as its Ethic guidelines and Sustainability model. I'm also happy that the author can submit a course in Spanish and that, as a reviewer, I'm allowed to add issues/comments in Spanish.
