### 2022-05-26 jaime-pru/Analisis-de-datos-espaciales#4 @hernandezcortes
My following comments are related to the introduction:

1. The introduction not only covers information on the basic syntax of R but ratherprovides information of vectors, matrices, and factor variables. The “Factores” subsection is missing a brief introduction of what a variable is and how it can be useful.

2. It might be nice to tell the reader that some of the functionalities of R come from baseR and others from specific packages. Readers would benefit from learning that distinction.

3. I particularly enjoyed how the author walked through the example of ggplot and what each part of the elements.

4. It would be nice to include a list of the learning objectives in the first part of the chapter. Given that the chapter covers a lot of material, the reader would be able to learn the contents and objectives from the very beginning.
