### 2022-05-26 jaime-pru/Analisis-de-datos-espaciales#9 @hernandezcortes
Comments on Chapter 2: Mapas coropléticos en R
Change “apuntamos” to “señalamos”.

Raster data need not to come from satellite images, it can come from other sources.

Not all spatial data have a time component, for example shapefiles might not necessarily have a time component. Perhaps clarify that raster data has three components whereas other spatial data might only have location and attributes.

Line 22: Change “El material aquí desarrollado se sirve de información espacial de tipo vectorial”. To “El material presentado utiliza información espacial de tipo vectorial”.

Line 92: “De la mañana de información”. This sentence is not clear.

It would be good to explain what coordinate systems and projections are. This chapter goes over st_crs without a lot of detail and I have found this to be an important aspect of any data analysis. Please add a couple of lines explaining what a coordinate system is, why it is important to have spatial data in the same coordinate system and how it might be possible to change it. This would allow the readers to understand the importance of coordinate systems and how they play an important role while plotting maps.

Explain in which circumstances it is useful to calculate the centroids and how it might be used in the future.

Subsection “Mapa de percentiles”: It would be helpful to explain why the user might want to plot percentiles map and why it is a nice way of plotting the data.
