### 2022-05-26 jaime-pru/Analisis-de-datos-espaciales#11 @hernandezcortes
Comments on Chapter #4: Modelos de regresión lineal
1. Line 9: please rewrite “Entre los tipos de preguntas planteadas se halla uno que es de particular interés” This sentence is not clear.

2. Line 10: the examples provided are related to macroeconomics. Perhaps you could add other examples that the readers could relate. For example: does higher education attainment causes higher salaries?

3. Line 28: please correct the (…)

4. Line 28: “mínimos cuadrados ordinarios”.

5. Line 30: “fórmula” instead of “formula”

6. Line 39: “minimice el error” not “minimice los errores”.

7. Line 91: Perhaps explain that this type of regression has one dependent variable and one independent variable only. This would allow the reader to distinguish between univariate regression and multivariate regression.

8. Line 101: the last part is missing a “)”.

9. In the discussion of linearity of the regression in terms of parameters, it would be useful to explain this further. I have found that students do not necessarily understand the importance of this assumption. Perhaps you can show two examples: one with an independent variable is linear and one with an independent variable that is quadratic, logarithmic or exponential and ask the reader which one is linear in parameters. The answer would be that both of them are but this will allow them to understand that the linearity assumption corresponds to the parameters and not the independent variables.

10. The chapter would benefit from explaining that the assumptions behind linear regression might not necessarily hold and explain examples where they don’t hold.

11. It would be good to do a brief explanation of what null and alternative hypotheses are.

12. My biggest concern with this chapter is that the author started the motivation of the chapter with an explanation about causality and how the theory of causality relates to Economics. However, after explaining what a regression is and how to perform it in R, there is no longer mention of causality. It would be good to: 1) explain clearly under which conditions regressions allow us to achieve causality, 2) explain that the correlation of two variables might not mean causality, and 3) the importance of using theory and regression analysis when thinking about causal questions.
