### 2022-05-26 jaime-pru/Analisis-de-datos-espaciales#12 @hernandezcortes
Comments on Chapter 5: Análisis espacial II: modelos econométricos espaciales
1. Line 3: “Mencionamos dos razones por las que” instead of “Mencionamos que dos son las razones por las que”.

2. Line 4: “fenómeno de interés no se” instead of “fenómeno de interes de no se”

3. I think that there is a big jump between chapter 4 and chapter 5. It might be good to remind the reader how regression analysis inform the concepts in this chapter.

4. Perhaps it would be good to explain nested models a bit more so that the reader understand its use and relevance.

5. The methodology to estimate the spatial econometric model could be explained a bit better. I suggest the author to add a few sentences on how to perform steps i) to iv), particularly steps ii), iii), and iv).

6. It would be nice to add a bit of intuition about the differences between modelos de rezago and modelos de error. Is there a way to judge this analytically or is this theory-driven?
