### 2022-05-26 #173 @hernandezcortes
I have completed the checklist provided and reviewed the JOSE paper and the course. The course is very well structured and relevant. I think that this course fills the existing gap of materials to learn R that are available in Spanish. The course will be relevant for students and professors in Spanish-speaking countries, which means that it can have a big impact for other users. Importantly, the author does not assume any previous knowledge from the students, which makes the course easy to follow for first-time R users. I have a few suggestions and recommendations below that could help the exposition of the material.

Please find below my general comments:

**General checks:**
1. I don’t think that the author has a Version but I could have missed this information.

**Documentation:**
1. It might be good to add that despite that the course covers several topics in the Economics field, the course could be good for any student that would like to learn about data analysis and how to effectively visualize it in R.

**Learning objectives:**
1. The course would benefit from a longer explanation of the objectives of the course and the tools that the students would acquire in simple terms.

**JOSE paper:**

The paper is well-written and motivated. I have a couple of minor suggestions:
1. “Economics” instead of “economics”.
2. I think that the course would be helpful not only for Economics students but rather any undergraduate student interested in learning quantitative methods. This would improve the impact of the course and the potential number of users.
3. “Meanwhile, chapter 4 presents a review of simple regression models” instead of “Meanwhile, in chapter 4, a very synthetic review of simple regression models is presented”.

**Overall:**
Chapters 1-4 are easy to follow and very well structured. I have a few minor suggestions that would improve the ability of the author to present the materials and convey the course message. Chapter 5 is a bit technical and I think that there was a big jump between chapter 4 and chapter 5. It might be good to remind the students the limitations of regression analysis and why spatial analysis is a powerful tool. Chapter 5 would also benefit from explaining the models with more intuition and examples.

**Specific comments to the chapters can be found below:**
Chapter 1: [https://github.com/jaime-pru/Analisis-de-datos-espaciales/issues/10](url)
Chapter 2: [https://github.com/jaime-pru/Analisis-de-datos-espaciales/issues/9](url)
Chapter 4: [https://github.com/jaime-pru/Analisis-de-datos-espaciales/issues/11](url)
Chapter 5: [https://github.com/jaime-pru/Analisis-de-datos-espaciales/issues/12](url)

Please let me know if there are any additional questions or concerns about my review. I really enjoyed reading the book and learned a lot from the author. I am looking forward to reading the final version of the book and learning more about spatial econometrics. Thank you so much for the opportunity to serve as a reviewer and I apologize for the late review.
