### 2022-11-08 3180031
Se antendió el comentarió sobre la ubicación de la universidad.
