+++ index.Rmd
@@ -43,7 +43,7 @@ Las ciencias regionales no son ya una disciplina menor, pasaron de ser una suert
 
 Si bien las ciencias regionales abarcan múltiples líneas temáticas, hay dos de ellas que se constituyen en tópicos clásicos y son lo que podría ser denominado _la corriente principal de las ciencias regionales_: el estudio de las fuerzas de aglomeración y los determinantes de la localización de la actividad. Buena parte de las investigaciones de estos ejes buscan responder a preguntas como ¿por qué las actividades económicas no se distribuyen de forma homogénea en el espacio?, o bien, ¿qué hace que determinadas actividades se localicen en unos sitios y no en otros?
 
-En su *Handbook of Regional Science*, Manfred Fischer y Peter Nijkamp señalan que quizá sean dos los elementos clave que constituyen a la ciencia regional: un enfoque multidisciplinario y un fuerte análisis cuantitativo. Respecto al primer elemento, si bien es cierto que la economía espacial, urbana y regional, han representado pasos alentadores hacia la aproximación del economista con otros científicos sociales, aquellos no suelen mirar hacia otras disciplinas con la frecuencia que exigen las problemáticas sociales. De este modo, aún sigue siendo necesario construir alternativas de formación en la fase final de los estudios de licenciatura que busquen romper la endogamia de la profesión. Esto es en alguna medida lo que en el [*Área de Concentración de Economía de la Innovación: empresas, redes y territorio*](https://rpm-uam.blogspot.com/p/docencia.html), que forma parte de la última fase de estudios en la licenciatura en Economía de la Universidad Autónoma Metropolitana Unidad Azcapotzalco, se ha buscado impulsar. 
+En su *Handbook of Regional Science*, Manfred Fischer y Peter Nijkamp señalan que quizá sean dos los elementos clave que constituyen a la ciencia regional: un enfoque multidisciplinario y un fuerte análisis cuantitativo. Respecto al primer elemento, si bien es cierto que la economía espacial, urbana y regional, han representado pasos alentadores hacia la aproximación del economista con otros científicos sociales, aquellos no suelen mirar hacia otras disciplinas con la frecuencia que exigen las problemáticas sociales. De este modo, aún sigue siendo necesario construir alternativas de formación en la fase final de los estudios de licenciatura que busquen romper la endogamia de la profesión. Esto es en alguna medida lo que en el [*Área de Concentración de Economía de la Innovación: empresas, redes y territorio*](https://rpm-uam.blogspot.com/p/docencia.html), que forma parte de la última fase de estudios en la licenciatura en Economía de la Universidad Autónoma Metropolitana Unidad Azcapotzalco en la Ciudad de México, México, se ha buscado impulsar. 
 
 En tanto, sobre el enfoque cuantitativo que caracteriza a las ciencias regionales, es cierto que en todos los planes de estudio de economía encontramos un sólido repertorio de instrumentos de carácter cuantitativo: matemáticas, estadística y, por supuesto, econometría. No obstante, las herramientas requeridas para el análisis de la realidad desde una perspectiva espacial siguen siendo escasas dentro de la formación a nivel de licenciatura. Estas notas buscan ser una contribución, por mínima que esta sea, para subsanar dicha situación.
 
