### 2022-11-11 jaime-pru/Analisis-de-datos-espaciales#1 @jaime-pru
In the corrections, instructions are given in Spanish and English. Besides, the instructions are more detailed.
