+++ _book/reference-keys.txt
@@ -1,94 +1,0 @@
-fig:unnamed-chunk-2
-fig:unnamed-chunk-3
-fig:unnamed-chunk-31
-fig:unnamed-chunk-110
-fig:unnamed-chunk-111
-fig:unnamed-chunk-112
-fig:unnamed-chunk-128
-fig:unnamed-chunk-154
-fig:unnamed-chunk-158
-fig:unnamed-chunk-172
-fig:unnamed-chunk-173
-fig:unnamed-chunk-174
-r-una-introducción-desde-la-exploración-de-información
-r-lenguaje-de-programación-y-plataforma-de-análisis
-códigos-objetos-paquetes
-código
-objetos-operadores-y-tipos-de-variables
-vectores
-matrices
-factores
-data-frame-o-arreglo-de-datos
-paquetes
-solicitar-ayuda-de-un-paquete-o-función-de-r
-cargar-bases-de-datos
-cargar-un-archivo-de-excel-xls-xslx
-cargar-archivos-separados-por-comas-csv
-cargar-bases-con-la-funcionalidad-importar
-exploración-inicial-a-través-de-gráficos
-manipulación-de-la-información
-seleccionar-variables
-filtrar-seleccionar-casos
-ordenar
-crear-nuevas-variables
-resúmenes-de-información-y-grupos
-mapas-coropléticos-en-r
-datos-y-datos-espaciales
-los-paquetes
-carga-de-la-base-de-datos-geográfica
-mapas-coropléticos-básicos
-personalización
-argumentos-de-personalización-dentro-de-la-función-de-relleno
-paleta-de-colores
-leyenda
-mapas-de-clasificación
-argumentos-de-personalización-del-borde
-función-para-elementos-de-diseño-de-salida-tm_layout
-mapa-base-interactivo
-cartograma
-centroides-o-coordenadas-geométricas-punto-medio
-tópico-adicional-mapas-de-clasificación-especial-elementos-de-programación
-una-palabra-de-advertencia
-mapa-de-intervalos-personalizados
-mapas-de-valores-extremos
-mapa-de-percentiles
-construcción-de-una-función-personalizada-para-hacer-mapas-de-percentiles
-análisis-espacial-i-autocorrelación
-autocorrelación-y-dependencia-espacial
-matrices-de-pesos-espaciales-en-r
-los-paquetes-1
-matrices-de-contigüidad-el-argumento-queen-de-la-función-poly2nb
-estadísticos-de-correlación-espacial
-la-matriz-de-pesos-espaciales-estandarizada
-el-rezago-o-retardo-espacial
-coeficiente-de-correlación-espacial-la-i-de-moran
-diagrama-de-dispersión-de-moran
-múltiples-elementos-de-personalización-de-ésta-y-otras-gráficas-asociadas-al-paquete-base-de-r-pueden-revisarse-en-la-documentación-de-la-función-par.
-índice-de-moran-local-y-mapa-de-clusters
-apéndice-un-enfoque-adicional-para-construir-matrices-de-pesos-espaciales
-matriz-tipo-torre-rook
-matriz-tipo-reyna-queen
-matrices-de-contigüidad-de-orden-superior
-visualización-de-la-estructura-espacial
-matrices-de-distancia-k-vecinos-más-cercanos
-modelos-de-regresión-lineal
-causalidad-y-correlación-el-alcance-del-análisis-de-regresión
-lo-artesanal-de-la-econometría
-qué-buscamos-a-través-de-una-regresión-con-mínimos-cuadrados-ordinarios
-modelo-de-regresión-lineal-simple
-supuestos-del-módelo-clásico-de-regresión-lineal
-regresión-lineal-simple-en-r
-una-exploración-visual-de-los-errores
[…]
