### 2022-11-15 4f004bf
actualización  de la justificación. Generaba problemas a un nivel h1 pero no al h2, lo que redujo el numero de alineamientos que habia que poner
