+++ 03-analisis-espacial-I.Rmd
@@ -65,9 +65,9 @@ Este tipo de matriz, denotado por la letra mayúscula $W$, contiene como element
 Aquí acabamos de ilustrar la lógica con la que puede ser construida una matriz de pesos espaciales a partir de una retícula regular con apenas seis elementos. Veamos ahora cómo obtener matrices de pesos espaciales sirviéndonos de R, ya que desarrollar los pasos anteriores para un vecindario compuesto por 76 elementos como los que integran el Valle de México es tarea para una máquina, no para nosotros.
 
 ## Matrices de pesos espaciales en R
-
-### Los paquetes
 <div style="text-align: justify">
+### Los paquetes
+
 En R hay muchas rutas para desarrollar la misma tarea. Lo que presentamos en este capítulo es una de múltiples maneras que pueden encontrarse para la construcción de matrices de pesos espaciales en R. En el apéndice, encontrarás otra alternativa, un tanto más elaborada. La manera en que desarrollamos en esta sección la tarea de construir las matrices de pesos espaciales se sirve de los siguientes paquetes:
 
 * `rgdal`: según se lee en la documentación, "proporciona enlaces a la biblioteca de abstracción de datos 'geoespaciales'"
@@ -87,7 +87,7 @@ library(rgdal)
 ```
 
 ### Matrices de contigüidad: el argumento `queen` de la función `poly2nb()`
-<div style="text-align: justify">
+
 Primero, carguemos la base de datos espacial. En este caso, recurrimos a la función `readOGR` que forma parte del paquete `rgdal` para leer los archivos de tipo SHP.
 
 ```{r}
@@ -173,9 +173,9 @@ plot(mFALSE, coordinates(covid_zmvm), add=TRUE, col='lightgreen')
 Como puedes observar, son estructuras muy parecidas, aunque es posible notar sus diferencias. Es ampliamente recomendable que revises el apéndice de este capítulo pues en él se trata un enfoque diferente para la construcción de matrices de pesos espaciales que, si bien es más complejo, te puede servir para avanzar en tu conocimiento sobre R.
 
 ## Estadísticos de correlación espacial
-
-### La matriz de pesos espaciales estandarizada
 <div style="text-align: justify">
+### La matriz de pesos espaciales estandarizada
+
 Antes dijimos que cada uno de los elementos $w_{ij}$  de la matriz $W$ pueden tomar como valores ceros o unos. La matriz $W$ puede escribirse como:
 
 
@@ -216,7 +216,7 @@ ii. ¿Logras identificar cómo la suma de cada renglón de la lista es igual a u
 ------------------------------------------------------------------------
 
 ### El rezago o retardo espacial
-<div style="text-align: justify">
+
 Regresemos a nuestra definición sobre la autocorrelación espacial: relación funcional existente entre los valores que adopta un indicador en una zona del espacio y en zonas vecinas. Ya hemos visto cómo podemos operacionalizar el concepto de vecindad, por lo que podemos ahora concentrémonos en los otros elementos. Comencemos con aquel que define el valor de una variable en las zonas vecinas, a lo que llamaremos *variable espacialmente rezagada*. Una variable espacialmente rezagada no es otra cosa que un promedio ponderado del valor de una variable en un sitio y sus locaciones vecinas [@Chasco2003,p. 61; @Anselin2020]. Siguiendo a [@Anselin2020], "el rezago espacial de $y$ del objeto espacial $i$ es expresado como $Wy_{i}$:
 
 $$
@@ -282,7 +282,7 @@ Los valores previamente listados corresponden tanto al nombre como al número de
 
 
 ### Coeficiente de correlación espacial: la I de Moran
-<div style="text-align: justify">
+
[…]
