+++ docs/reference-keys.txt
@@ -1,15 +1,14 @@
 fig:unnamed-chunk-2
-fig:unnamed-chunk-3
-fig:unnamed-chunk-31
+fig:unnamed-chunk-30
+fig:unnamed-chunk-109
 fig:unnamed-chunk-110
 fig:unnamed-chunk-111
-fig:unnamed-chunk-112
-fig:unnamed-chunk-128
-fig:unnamed-chunk-154
-fig:unnamed-chunk-158
+fig:unnamed-chunk-127
+fig:unnamed-chunk-153
+fig:unnamed-chunk-157
+fig:unnamed-chunk-171
 fig:unnamed-chunk-172
 fig:unnamed-chunk-173
-fig:unnamed-chunk-174
 r-una-introducción-desde-la-exploración-de-información
 r-lenguaje-de-programación-y-plataforma-de-análisis
 códigos-objetos-paquetes
