+++ 01-intro.Rmd
@@ -16,9 +16,9 @@ A lo largo de este libro usaremos R a través de [*RStudio*](https://es.wikipedi
 
 R puede ser descargado de [aquí](https://www.r-project.org/). Al seleccionar *"download R"* hay que elegir cualquier CRAN Mirror (del inglés *Comprehensive R Archive Network* que son repositorios que contienen una copia del código base de R y sus paquetes). México tiene dos, puedes elegir cualquiera.
 
-En tanto, R Studio puede ser desargado de [aquí](https://www.rstudio.com/). Ve a la sección *"Download"* en la parte superior y elije la versión gratuita. El navegador debería seleccionar la versión adecuada para tu equipo, sino es así, selecciona del listado que aparece en la parte inferior. También puedes consultar [este video](https://youtu.be/zlnm1jLWLeE) para seguir paso a paso la instalación de ambos programas.
+En tanto, RStudio puede ser descargado de [aquí](https://www.rstudio.com/). Ve a la sección *"Download"* en la parte superior y elije la versión gratuita. El navegador debería seleccionar la versión adecuada para tu equipo, sino es así, selecciona del listado que aparece en la parte inferior. También puedes consultar [este video](https://youtu.be/zlnm1jLWLeE) para seguir paso a paso la instalación de ambos programas.
 
-Al abrir RStudio se mostrará una pantalla como la que aparece en la figura 1.1. La pantalla principal de RStudio está dividida en cuatro partes. En caso de que el recuadro superior izquierdo no aparezca, da clic en el ícono de la hoja en blanco con el signo de más (*New file*, parte superior izquierda de la pantalla) para que se despliegue.
+Al abrir RStudio se mostrará una pantalla como la que aparece en la Figura 1.1. La pantalla principal de RStudio está dividida en cuatro partes. En caso de que el recuadro superior izquierdo no aparezca, da clic en el ícono de la hoja en blanco con el signo de más (*New file*, parte superior izquierda de la pantalla) para que se despliegue.
 
 ```{r, echo=FALSE, fig.align='center',out.width='100%',fig.cap='Panel principal'} 
 knitr::include_graphics("recursos intro/Figura_1-Entorno.jpg")
@@ -41,27 +41,36 @@ En este sentido, se debe señalar que R es sensible al uso de mayúsculas y min�
 
 ### Objetos, operadores y tipos de variables
 
-Si bien R es tremendamente potente, puede incluso utilizarse como simple calculadora. Los operadores matemáticos que pueden emplearse en R se muestran en la figura 1.2:
+Si bien R es tremendamente potente, puede incluso utilizarse como simple calculadora. Los operadores matemáticos que pueden emplearse en R se muestran en el cuadro 1.1:
 
-```{r, echo=FALSE, fig.align='center',out.width='120%', fig.cap='Operadores'} 
-knitr::include_graphics("recursos intro/Figura_2-Operadores.jpg")
-```
 
-De esta forma sí colocas en la consola el número `5` seguido del signo aritmético correspondiente, por ejemplo un signo de más, `+`, y luego otro número, R devolverá el resultado de la operación:
+|Operador | Operación | Ejemplo | Resultado |
+|:-------:|:---------:|:-------:|:---------:|
+|+        |Suma       |5+3      |8          |
+|-        |Resta      |5-3      |2          |
+|*        | Multiplicación | 5*3|15         |
+|/        |División   |5/3      |1.66667    |
+|^        |Potencia   |5^3      |125        |
+|%%       |División entera|5%%3 |2          |
+
+:Cuadro 1.1
+
+De esta forma si colocas en la consola el número `5` seguido del signo aritmético correspondiente, por ejemplo un signo de más, `+`, y luego otro número, R devolverá el resultado de la operación:
 
 ```{r}
 5+10
 ```
 
[…]
