+++ index.Rmd
@@ -48,7 +48,7 @@ Manual de Ciencias Regionales), Manfred Fischer y Peter Nijkamp señalan que qui
 
 En tanto, sobre el enfoque cuantitativo que caracteriza a las ciencias regionales, es cierto que en todos los planes de estudio de economía encontramos un sólido repertorio de instrumentos de carácter cuantitativo: matemáticas, estadística y, por supuesto, econometría. No obstante, las herramientas requeridas para el análisis de la realidad desde una perspectiva espacial siguen siendo escasas dentro de la formación a nivel de licenciatura. Estas notas buscan ser una contribución, por mínima que esta sea, para subsanar dicha situación.
 
-Este trabajo está, por tanto, fundamentalmente dirigido a los estudiantes de licenciatura interesados en el desarrollo de habilidades técnicas para el análisis cuantitativo que exigen las ciencias regionales y el enfoque espacial de la economía. Así pues, el objetivo de estas notas es guiar al estudiante, desde un enfoque fundamentalmente práctico, en el conocimiento y manejo de técnicas para la exploración, análisis y modelado de información espacial mediante el uso de **R**, un programa informático y lenguaje de programación enfocado en el análisis estadístico y visualización de información y **RStudio** un entorno de desarrollo integrado (IDE) desde donde se puede interactuar con **R** más eficientemente.
+Este trabajo está, por tanto, fundamentalmente dirigido a los estudiantes de licenciatura interesados en el desarrollo de habilidades técnicas para el análisis cuantitativo que exigen las ciencias regionales y el enfoque espacial de la economía. Así pues, el objetivo de estas notas es guiar al estudiante, desde un enfoque fundamentalmente práctico, en el conocimiento y manejo de técnicas para la exploración, análisis y modelado de información espacial mediante el uso de **R**, un programa informático de software libre y lenguaje de programación enfocado en el análisis estadístico y visualización de información y **RStudio** un entorno de desarrollo integrado (IDE) desde donde se puede interactuar con **R** más eficientemente.
 
[…]
