### 2022-12-01 932c0fb
Cambios hasta el 30 de noviembre de 2022
