+++ 03-analisis-espacial-I.Rmd
@@ -1,16 +1,35 @@
 # Análisis espacial I: autocorrelación
+<div style="text-align: justify">
+Observa el siguiente mapa. En él, se representa la variable índice de de desarrollo humano que calculó el Programa de Naciones Unidad para el Desarrollo (PNUD) en 2015.
 
-## Autocorrelación y dependencia espacial
+```{r echo=FALSE, message=FALSE, warning=FALSE}
+library(tmap)
+library(sf)
+base <-sf::st_read("base de datos\\covid_zmvm shp\\covid_zmvm.shp")
+tmap::tm_shape(base) +
+  tmap::tm_borders()+
+  tmap::tm_fill("idh2015", 
+                style = "quantile", 
+                n=6,
+                title = "IDH 2015",
+                palette = "GnBu",
+                legend.format = list(text.separator="-",digits=2))+
+  tmap::tm_layout(main.title = "IDH, 2015", main.title.position = "center", title.size = 0.5)+
+  tm_text("nom_mun")
+
+```
+
+:Figura 3.1. Índice de desarrollo humano para los municipios de la ZMVM, 2015
+
+¿Notas algún patrón? ¿Detectas cómo los municipios con un valor alto del índice de desarrollo (Benito Juárez, Alvaro Obregón, Miguel Hidalgo, Coyoacán y otros) se encuentran _agrupados_ en la parte centro-sur de la región considerada? Esto es un ejemplo de la posible presencia de *autocorrelación espacial*. Cuando tratamos información espacial hay altas probabilidades de que los valores de un indicador en el espacio tengan en las cercanías valores muy parecidos (o a veces, al contrario, nada parecidos). Esta característica de los datos georeferenciados hace que sean necesarias ciertas técnicas para su tratamiento. En este capítulo trataremos con el concepto clave de autocorrelación espacial y veremos a través de que instrumentos es posible medirla para, más adelante, incorporar su riqueza informativa una explicación sustantiva de los fenómenos socioterritoriales analizados.
+
+## Autocorrelación espacial y definición de vecindad
 <div style="text-align: justify">
 La identificación de autocorrelación espacial es importante como parte del proceso de análisis del fenómeno socioterritoriales al menos por dos cuestiones, una de carácter técnico y otra de carácter sustantivo. Respecto a los elementos técnicos es posible sostener que, si existe autocorrelación espacial en nuestros datos, lo más probable es que los métodos de estimación de los parámetros del modelo propuesto a través de mínimos cuadrados ordinarios dejen de ser válidos, en la medida en que no se cumplen los supuestos que requiere dicho procedimiento, específicamente, que los errores o perturbaciones del modelo no estén correlacionados. Por otro lado, cuando nos referimos a elementos sustantivos estamos queriendo expresar que nuestro fenómeno de interés, aquél que buscamos captar a través del modelo propuesto, en realidad estaría dejando de lado un comportamiento sistemático de asociación a nivel territorial.
 
-Sobre este segundo punto detengámonos un momento. ¿Qué puede estar ocurriendo que hace que los fenómenos se no se distribuyan aleatoriamente en el espacio? Por ejemplo, ¿por qué se concentra la actividad económica en determinadas ciudades o por qué algunos servicios sólo se brindan en una zona de la ciudad? Esto respecto a fenómenos económicos, pero ¿qué hay con el hecho de que una enfermedad se concentra notoriamente en algunas áreas de la ciudad y no en otras? Dicho en otras palabras, ¿qué hay detrás de la formación de un patrón en la forma en que se distribuye un fenómeno en el espacio y cómo puede ser explicado? A eso nos referimos cuando decimos que hay elementos sustantivos para el análisis al hallar evidencia de autocorrelación espacial.
[…]
