+++ book.bib
@@ -167,3 +167,29 @@ For this notebook, we will continue to use the socioeconomic data for 55 New Yor
    title = {Exploring Spatial Data with GeoDaTM : A Workbook},
    year = {2005},
 }
+
+@book_section{Nijkamp2015,
+   author = {Peter Nijkamp and Waldemar Ratajczak},
+   doi = {10.1007/978-3-319-07305-7_2},
+   journal = {Regional Science Matters: Studies Dedicated to Walter Isard},
+   title = {The spatial economy: A holistic perspective},
+   year = {2015},
+}
+
+@book{Asuad2001,
+   author = {Normand E. Asuad Sanén},
+   city = {Puebla},
+   edition = {1},
+   journal = {Science},
+   publisher = {Universidad Autónoma de Puebla},
+   title = {Economía regional y urbana: introducción a las teorías, técnicas y metodologías básicas},
+   year = {2001},
+}
+
+@book{Olaya2020,
+   author = {Victor Olaya},
+   isbn = {978-1-71677-766-0},
+   title = {Sistemas de Información Geográfica},
+   url = {https://volaya.github.io/libro-sig/index.html},
+   year = {2020},
+}
