### 2022-12-14 8f96f93
Cambios al capítulo de análisis espacial I.
