+++ 03-analisis-espacial-I - rgeoda.Rmd
@@ -1,0 +1,706 @@
+# Análisis espacial I: autocorrelación
+<div style="text-align: justify">
+Observa el siguiente mapa. En él, se representa la variable índice de de desarrollo humano que calculó el Programa de Naciones Unidad para el Desarrollo (PNUD) en 2015.
+
+```{r echo=FALSE, message=FALSE, warning=FALSE}
+library(tmap)
+library(sf)
+base <-sf::st_read("base de datos\\covid_zmvm shp\\covid_zmvm.shp")
+tmap::tm_shape(base) +
+  tmap::tm_borders()+
+  tmap::tm_fill("idh2015", 
+                style = "quantile", 
+                n=6,
+                title = "IDH 2015",
+                palette = "GnBu",
+                legend.format = list(text.separator="-",digits=2))+
+  tmap::tm_layout(main.title = "IDH, 2015", main.title.position = "center", title.size = 0.5)+
+  tm_text("nom_mun")
+
+```
+
+:Figura 3.1. Índice de desarrollo humano para los municipios de la ZMVM, 2015
+
+¿Notas algún patrón? ¿Detectas cómo los municipios con un valor alto del índice de desarrollo (Benito Juárez, Alvaro Obregón, Miguel Hidalgo, Coyoacán y otros) se encuentran _agrupados_ en la parte centro-sur de la región considerada? Dicho patrón ilustra la posible presencia de *autocorrelación espacial*, es decir, que los valores de un indicador en una zona estén rodeados por valores muy semejantes o nada semejantes: en ambos casos decimos que hay autocorrelación. Cuando tratamos con datos de una variable georeferenciada hay altas probabilidades de que dicha variable esté autocorrelacionada y esto hace necesario que sean necesarias ciertas técnicas para su tratamiento, es decir, para su representación y modelación. En este capítulo trataremos con el concepto clave de autocorrelación espacial y veremos a través de que instrumentos es posible medirla para, más adelante, incorporar su riqueza informativa a una explicación sustantiva de los fenómenos socioterritoriales analizados.
+
+## Autocorrelación espacial y definición de vecindad
+<div style="text-align: justify">
+
+Una definición sintética de autocorrelación espacial es la que nos brinda Chasco (2003) como “la relación funcional existente entre los valores que adopta un indicador en una zona del espacio y en zonas vecinas” [@Chasco2003,p. 49]. Por ejemplo, imagina que el barrio de la ciudad donde vives presenta un alto número de contagios por COVID19 y, además, los barrios vecinos tienen también valores altos: en este caso es probable que tengamos autocorrelación espacial positiva.
+
+La identificación de autocorrelación espacial es importante como parte del proceso de análisis del fenómeno socioterritoriales al menos por dos cuestiones, una de carácter técnica durante la modelación econométrica y otra de carácter sustantiva en relación con la explicación del fenómeno. Respecto a la razón técnica, si existe autocorrelación espacial en nuestros datos lo más probable es que la estimación de los parámetros de un modelo con de mínimos cuadrados ordinarios deje de ser válida, en la medida en que no se cumplen los supuestos que requiere dicho procedimiento, específicamente, que los errores o perturbaciones del modelo no estén correlacionados; sobre esto abundaremos en el capítulo siguiente cuando hagamos un repaso de los modelos clásicos de regresión lineal.  
+
[…]
