+++ 03-analisis-espacial-I.Rmd
@@ -282,10 +282,10 @@ mTRUE[[1]]#Devuelve los números de identificación de los municipios y alcaldí
 Los números anteriores corresponden a los vecinos de Álvaro Obregón, pero, ¿cómo saber su nombre y el número de casos positivos de cada uno? El siguiente segmento de código nos dará los nombres de los vecinos y el número de casos positivos en cada uno.
 
 ```{r}
-#Crea un objeto que contienw el nombre de las alcaldías y municipios vecinos de Álvaro Obregón
+#Crea un objeto que contiene el nombre de las alcaldías y municipios vecinos de Álvaro Obregón
 nom_vec <- covid_zmvm$nom_mun[mTRUE[[1]]]
 
-#Crea un objeto que contiene el valor de pos_hab para cada observación vecina de Álvaro Obregón
+#Crea un objeto que contiene el valor de la variable pos_hab (casos positivos por mil habitantes) para cada observación vecina de Álvaro Obregón
 covi_vec <- covid_zmvm$pos_hab[mTRUE[[1]]]
 
 # Guarda los dos objetos anteriores en un dataframe y lo muestra en una tabla
@@ -307,19 +307,21 @@ ii. Obtén un cuadro con las lista de vecinos y los valores de casos positivos p
 
 ### Coeficiente de correlación espacial: la I de Moran
 
-Ahora que ya tenemos todos los elementos que integran la definición de autocorrelación espacial, ¿cómo la medimos? Es decir, como sabemos si la variable casos positivos por COVID19 presenta o no dicha característica. Para ello, debemos construir una medida, un coeficiente, es decir, un estadístico.
+Una vez que hemos abordado y resuelto el problema de cómo definir vecindad a partir de la matriz de pesos espaciales, tenemos todos los elementos que integran la definición de autocorrelación espacial: relación funcional, valor de una variable y relación de vecindad. Ahora bien, ¿cómo la medimos? Es decir, cómo sabemos si la variable casos positivos por COVID19 está espacialmente correlacionada. Al principio de este capítulo dijimos que apreciar ciertos patrones de agrupamiento en un mapa, como el del índice de desarrollo humano, era un posible indicio de autocorrelación espacial. Es momento de formalizar dicho indicio a través de un indicador apropiado, debemos pues construir un estadístico, un estadístico de autocorrelación espacial.
 
-Sin duda, el estadístico de asociación espacial más común es la I de Moran, que no es otra cosa que un coeficiente de correlación lineal que "incorpora al espacio", es decir, mide la asociación lineal entre la variable de interés y su rezago espacial.
+El estadístico de asociación espacial más socorrido es el propuesto por [Patrick Moran](https://en.wikipedia.org/wiki/P._A._P._Moran) la I de Moran. La I de Moran es un coeficiente de correlación lineal que "incorpora al espacio", es decir, mide la asociación lineal entre una variable (digamos casos positivos) y su rezago espacial (el valor promedio de los casos positivos de los vecinos). Como todo coeficiente de asociación, el valor de la I de Moran se encuentra entre -1 y 1.  
 
-Para calcular el coeficiente o estadístico de Moran necesitamos recurrir a la función `moran.test()` del paquete `spdep`. Los argumentos de la función deben especificar el nombre de la variable y el tipo de estructura espacial dado por la matriz de pesos usada; adicionalmente, se puede indicar qué hacer en caso de que existan islas (objetos espaciales sin vecinos) con el argumento `zero.policy`.
[…]

[…]
