+++ book.bib
@@ -110,7 +110,6 @@ For this notebook, we will continue to use the socioeconomic data for 55 New Yor
 }
 
 @book{Montgomery2012,
-   abstract = {Regression analysis is one of the most widely used techniques for analyzing multifactor data. Its broad appeal and usefulness result from the conceptually logical process of using an equation to express the relationship between a variable of interest (the response) and a set of related predictor variables. Regression analysis is also interesting theoretically because of elegant underlying mathematics and a well - developed statistical theory. Successful use of regression requires an appreciation of both the theory and the practical problems that typically arise when the technique is employed with real - world data. This book is intended as a text for a basic course in regression analysis. It contains the standard topics for such courses and many of the newer ones as well. It blends both theory and application so that the reader will gain an understanding of the basic principles necessary to apply regression model - building techniques in a wide variety of application environments. The book began as an outgrowth of notes for a course in regression analysis taken by seniors and fi rst - year graduate students in various fi elds of engineering, the chemical and physical sciences, statistics, mathematics, and management. We have also used the material in many seminars and industrial short courses for professional audiences. We assume that the reader has taken a fi rst course in statistics and has familiarity with hypothesis tests and confi - dence intervals and the normal, t , ?? 2 , and F distributions. Some knowledge of matrix algebra is also necessary. The computer plays a signifi cant role in the modern application of regression. Today even spreadsheet software has the capability to fi t regression equations by least squares. Consequently, we have integrated many aspects of computer usage into the text, including displays of both tabular and graphical output, and general discussions of capabilities of some software packages. We use Minitab�, JMP�, SAS�, and R for various problems and examples in the text. We selected these packages because they are widely used both in practice and in teaching regression and they have good regression. Many of the homework problems require software for their solution. All data sets in the book are available in electronic form from the publisher. The ftp site ftp://ftp.wiley.com/public/sci_tech_med/introduction_linear_ regression hosts the data, problem solutions, PowerPoint fi les, and other material related to the book.},
    author = {Montgomery. Douglas C. and Peck. Elizabeth A. and Vining. G. Geoffrey},
    issn = {03067734},
    issue = {2},
@@ -120,14 +119,6 @@ For this notebook, we will continue to use the socioeconomic data for 55 New Yor
    year = {2012},
 }
 
-@webpage{JHU2020,
-   author = {John Hopkins University},
-   journal = {Data scientit's tool box},
-   month = {4},
-   title = {Types of data science questions},
-   year = {2020},
-}
-
 @webpage{Dalpiaz2022,
    author = {David Dalpiaz},
    title = {Applied Statistics with R},
@@ -135,10 +126,9 @@ For this notebook, we will continue to use the socioeconomic data for 55 New Yor
 }
 
 @book{Gujarati2011,
-   abstract = {This book covers some essential topics of econometrics. It covers from single regression to multiple regression. The second part of the book talks about how to detect a violation of assumptions (multicollinearity, heteroscedasticity, autocorrelation, model specification) made for running multiple regression and what the remedies are. The third part deals with three topics, including (a) regression on dummy variables, (b) regression on dummy dependent variables, (c) autoregressive and distributed lag models. The last part deals with simultaneous-equation model.},
    author = {Damodar Gujarati and N Dawn C. Porter},
    journal = {Basic Econometrics},
-   title = {Econometria Basica},
+   title = {Econometria Básica},
    year = {2011},
 }
 
@@ -192,4 +182,11 @@ For this notebook, we will continue to use the socioeconomic data for 55 New Yor
    title = {Sistemas de Información Geográfica},
[…]
