+++ 04-regresion-lineal.Rmd
@@ -1,19 +1,24 @@
 # Modelos de regresión lineal
+<div style="text-align: justify">
+Hasta este punto hemos hecho un recorrido que va desde los elementos más básicos sobre la exploración de información (capítulo 1), pasando por la representación de información espacial a través de mapas coropléticos (capítulo 2), hasta la exploración de patrones de asociación de una variable en el espacio, la llamada autocorrelación (capítulo 3).  
 
+Antes de avanzar en el conocimiento de las herramientas que nos permitirán elaborar algunos modelos básicos que incorporen los patrones espaciales que hemos identificado a través de econometría espacial (capítulo 5), en este capítulo hacemos un pequeño paréntesis en el tratamiento de la información espacial para llevar a cabo un repaso elemental, desde una perspectiva práctica, del modelo clásico de regresión lineal. Esto nos permitirá que tengas en la mente dos cosas: i) qué es lo que se busca a través de modelos econométricos y ii) cuáles son los supuestos detrás del modelo clásico de regresión lineal estimado con el método de Mínimos Cuadrados Ordinarios (MCO). 
 
 ## Causalidad y correlación: el alcance del análisis de regresión
-<div style="text-align: justify">
-En el proceso de investigación podemos plantearnos preguntas con un grado de complejidad diverso. Habrá preguntas que, al ser respondidas, nos permitan tener una descripción general del fenómeno, habrá otras que nos permitan explorar la relación entre dos fenómenos, o bien, preguntas cuya respuesta nos permita predecir el comportamiento futuro del elemento estudiado. Así pues, hay [diferentes tipos de preguntas](https://bit.ly/DatSciQ), dependiendo de su grado de complejidad. 
 
-Entre los tipos de preguntas planteadas se halla uno que es de particular interés: el análisis causal. El objetivo del análisis causal es "ver qué sucede con una variable cuando manipulamos otra variable -mirar la causa y el efecto de una relación"[@JHU2020]. Con el análisis causal se busca saber si un par de variables están asociadas, más aún, si si el cambio en una variable causa cambios en la otra. Responder este tipo de preguntas exige el dominio de diversas herramientas estadísticas, cuyo abordaje está completamente fuera del alcance de este capítulo. No obstante, el anális de regresión nos brinda una oportunidad que, cuando es bien utilizada, nos permite aproximarnos al análisis causal.
+En el proceso de investigación podemos plantearnos preguntas con un grado de complejidad diverso. Habrá preguntas que, al ser respondidas, nos permitan tener una descripción general del fenómeno, habrá otras que nos permitan explorar la relación entre dos fenómenos, o bien, preguntas cuya respuesta nos permita predecir el comportamiento futuro del elemento estudiado. Así pues, hay [diferentes tipos de preguntas de investigación](https://bookdown.org/rdpeng/artofdatascience/types-of-questions.html), dependiendo de su grado de complejidad. En su libro "El arte de la ciencia de datos", Roger Peng y Elizabeth Matsui[-@PengMatzui2017], abordan sistemáticamente esta cuestión.   
 
-La regresión como instrumento analítico nos brinda los elementos para acercarnos a la comprensión de la *asociación entre dos variables*, aunque no es posible hablar en sentido estricto de causalidad como resultado del análisis efectuado. Así, "El análisis de regresión solo puede abordar los problemas de correlación. No puede abordar el problema de la necesidad (causalidad). Por tanto, nuestras expectativas de descubrir relaciones de causa y efecto a partir de la regresión deberían ser modestas" [@Montgomery2012, p.3].
[…]
