+++ _main.Rmd
@@ -1,185 +1,0 @@
---- 
-title: "Análisis de datos espaciales con R"
-author: "Jaime Alberto Prudencio Vázquez"
-date: "`r Sys.Date()`"
-knit: "bookdown::render_book"
-site: bookdown::bookdown_site
-output:
- bookdown::bs4_book:
-  theme:
-    primary: "#3860b6"
-    base_font: "Segoe UI"
-  
-documentclass: book
-bibliography: [book.bib]
-fontsize: 11pt
-fontfamily: times
-classoption: oneside
-openright: yes
-geometry: 
-- tmargin=2.5cm
-- bmargin=2.5cm
-- lmargin=2.5cm
-- rmargin=2.5cm
-link-citations: yes
-header-includes:
-  - \usepackage[utf8]{inputenc} #para usar accentos
-  - \usepackage[Glenn]{fncychap} #Formato de titulo de capitulo
-  - \pagestyle{plain}
-lang: es
----
-
-
-
-<!--chapter:end:index.Rmd-->
-
-
-# R: Una introducción desde la exploración de información
-
-Placeholder
-
-
-## R: Lenguaje de programación y plataforma de análisis
-## Códigos, objetos, paquetes
-### Código
-### Objetos, operadores y tipos de variables
-#### Vectores
-#### Matrices
-#### Factores
-#### _Data frame_ o arreglo de datos
-### Paquetes
-### Solicitar ayuda de un paquete o función de R
-## Cargar bases de datos
-### Cargar un archivo de Excel (xls, xslx)
-### Cargar archivos separados por comas (csv)
-### Cargar bases con la funcionalidad importar
-## Exploración inicial a través de gráficos
-## Manipulación de la información
-### Seleccionar variables
-### Filtrar (seleccionar casos)
-### Ordenar
-### Crear nuevas variables
-### Resúmenes de información y grupos
-
-<!--chapter:end:01-intro.Rmd-->
-
-
-# Mapas coropléticos en R
-
-Placeholder
-
-
-## Datos y datos espaciales
-## Los paquetes
-## Carga de la base de datos geográfica
-## Mapas coropléticos básicos
-## Personalización
-### Argumentos de personalización dentro de la función de relleno
-#### Paleta de colores
-#### Leyenda
-#### Métodos de clasificación en los mápas
-### Argumentos de personalización del borde
-## Función para elementos de diseño de salida, `tm_layout()`
-## Mapa base interactivo
-## Cartograma
-## Centroides o coordenadas geométricas (punto medio)
-## Tópico adicional: Mapas de clasificación especial (elementos de programación)
-### Una palabra de advertencia
-### Mapa de intervalos personalizados
-### Mapas de valores extremos
-#### Mapa de percentiles
-### Construcción de una función personalizada para hacer mapas de percentiles
-
-<!--chapter:end:02-Mapas.Rmd-->
-
-
-# Análisis espacial I: autocorrelación
-
-Placeholder
-
-
-## Autocorrelación espacial y definición de vecindad
-## Matrices de pesos espaciales en R
-### Los paquetes
-### Matrices de contigüidad: el argumento `queen` de la función `poly2nb()`
-## Estadísticos de correlación espacial
-### La matriz de pesos espaciales estandarizada
-### El rezago o retardo espacial
-### Coeficiente de correlación espacial: la I de Moran
-### Diagrama de dispersión de Moran
-## Índice de Moran local y mapa de clusters
-## Apéndice: Un enfoque adicional para construir matrices de pesos espaciales
-### Matriz tipo torre (*rook*)
-### Matriz tipo reyna (*queen*)
-### Matrices de contigüidad de orden superior
[…]

[…]
