+++ 05-analisis-espacial-II.Rmd
@@ -1,13 +1,23 @@
 # Análisis espacial II: modelos econométricos espaciales
 
 <div style="text-align: justify">
-En el capítulo 3 nos hemos referido a la necesidad de identificar la presencia de autocorrelación espacial en el conjunto de información utilizado. Mencionamos que dos son las razones por las que conviene saber si la información usada presenta este rasgo: una técnica y otra sustantiva. En términos de la segunda, la presencia de autocorrelación espacial significa que el fenómeno de interés de no se distribuye de manera aleatoria en el espacio por lo que *algo* está pasando que debe ser investigado. En términos de la razón de carácter técnico, que no está desvinculada necesariamente a la razón previa, estimar un modelo clásico de regresión lineal a través del método usual de Mínimos Cuadrados Ordinarios implicaría violar uno de sus supuestos, específicamente, aquel que tiene que ver con la independencia de los términos de error. Por tanto, por razones de la técnica utilizada y del interés de investigador, es necesario recurrir a una alternativa de modelación que permita tanto resolver el problema de la violación del citado supuesto como poner a disposición del interesado información relevante para buscar las causas que explican la distribución no aleatoria del fenómeno en el espacio.
+En el capítulo 3 nos hemos referido a la necesidad de identificar la presencia de autocorrelación espacial en el conjunto de información utilizado. Mencionamos que dos son las razones por las que conviene saber si la información usada presenta este rasgo: una de carácter técnico y otra que llamamos sustantiva. En términos de la segunda, la presencia de autocorrelación espacial significa que el fenómeno de interés no se distribuye de manera aleatoria en el espacio por lo que *algo está pasando y debe ser investigado*, ¿por qué la variable exhibe patrones en su distribución en el espacio? Por otro lado, en términos de la razón de carácter técnico, que no está desvinculada necesariamente de la razón previa, al estimar un modelo clásico de regresión lineal usando datos espaciales a través del método de Mínimos Cuadrados Ordinarios seguramente se violaría uno de sus supuestos, específicamente, aquel que tiene que ver con la independencia de los términos de error. Por tanto, por razones de la técnica utilizada y del interés de la investigadora, es necesario recurrir a un modelo econométrico espacial. Al hacerlo, se puede resolver el problema de la violación del citado supuesto y a la vez brindar elementos para buscar las causas que explican la formación de patrones en el espacio.
 
-En este capítulo nos dedicaremos a mostrar dos de las alternativas para lograr lo anterior a través de los llamados modelos econométricos espaciales. Es crucial que, antes de abordar este capítulo, hayan quedado plenamente asimilados los contenidos del capítulo 3 y 4, en la medida en que lo expuesto aquí se basa en los conocimientos previos.
+En este capítulo nos dedicamos a mostrar dos de las alternativas de modelos econométricos espaciales, es decir, modelos econométricos que incorporan diversos tipos de interacciones que se suceden en el espacio. Es crucial que, antes de abordar este capítulo, hayas asimilado los contenidos del capítulo 3 y 4, en la medida en que lo expuesto aquí se basa en los conocimientos previos.  
 
-## Diversas alternativas de modelos econométricos
+## Econometría espacial
 
[…]
