+++ book.bib
@@ -7,6 +7,37 @@
    year = {2014},
 }
 
+@book{Quintana2016,
+   author = {Luis Quintana and Miguel Ángel Mendoza},
+   city = {Ciudad de México},
+   journal = {Econometría Aplicada Utilizando Regresiones},
+   publisher = {Universidad Nacional Autónoma de México, DGAPA-FES Acatlán},
+   title = {Econometría aplicada usando R},
+   year = {2016},
+}
+
+@article{Cliff1981,
+   author = {A. D. Cliff and P. Haggett and J. K. Ord and G. R. Versey},
+   doi = {10.1016/0305-7488(82)90016-0},
+   issn = {03057488},
+   journal = {Spatial diffusion: an historical geography of epidemics in an island community ( iceland).},
+   title = {Spatial diffusion: an historical geography of epidemics in an island community ( iceland).},
+   year = {1981},
+}
+
+@article{Rezaeian2007,
+   author = {Mohsen Rezaeian and Graham Dunn and Selwyn St. Leger and Louis Appleby},
+   doi = {10.1136/jech.2005.043117},
+   issn = {0143005X},
+   issue = {2},
+   journal = {Journal of Epidemiology and Community Health},
+   title = {Geographical epidemiology, spatial analysis and geographical information systems: A multidisciplinary glossary},
+   volume = {61},
+   year = {2007},
+}
+
+
+
 @article{Chasco2003,
    abstract = {El objetivo fundamental de esta tesis consiste en proponer la predicción\nde datos espaciales como parte de la econometría espacial, a partir de\nuna metdología fundamentada en instrumentos exploratorios y confirmatorios\npropios de esta disciplina, En concreto, se desarrolla una subdisciplina\nde la predicción espacial, denominada predicción-extrapolación espacial,\ndistinta del método geoestadístico de interpolación espacial, entendida\ncomo el conjunto de métodos inferenciales destinados a obtener datos microterritoriales\na partir de información espacial agregada.\n\nLa consecución de este objetivo general exige una presentación y sistematización\ndel conjunto de técnicas de análisis exploratorio y confirmatorio de la\neconometría espacial, realizada en el cuerpo central de esta obra, con\nla ayuda de un capítulo completo y actualizando de Bibliografía, así como\nde unos Anexos, que amplían algunas ideas sin romper el hilo de la exposición.\n\nLa metodología propuesta se evalúa a través de un ejercicio de predicción-extrapolación\nde la renta familiar disponible de los municipios de la Comundiad de Madrid,\nen el que se hacen confluir tanto la utilización de técnicas de econometría\nespacial (a través de un "software" adecuado), como la experiencia de la\nautora en análisis microterritoriales, como Directora A. Del Área de Economía\nEspacial Microterritorial del Instituto Lawrence R. Klein (Universidad\nAutónoma de Madrid), co-dirigido por D. José Vicéns Otero, director de\nesta tesis.\n\nDado el interés y el cierto desconocimiento que aún pesa sobre el tema\nde la econometría espacial, esta tesis se plantea también, como objetivo,\nser de utilidad para todos los investigadores sociales interesados en el\nanálisis territorial, en cualquiera de sus formas, así como contribuir,\nen la medida de los posible, a la difusión de las técnicas de econometría\nespacial en nuestro país, sobre todo en el ámbito de los estudios},
    author = {Coro Chasco},
@@ -37,7 +68,6 @@
 }
 
 @book{Wickham2016,
[…]
