+++ paper.bib
@@ -16,6 +16,85 @@
    year = {2015},
 }
 
+@book{Fujita2000,
+   author = {Masahisa Fujita and Paul Krugman and Anthony J. Venables},
+   city = {Barcelona},
+   publisher = {Ariel},
+   title = {Economia espacial : las ciudades, las regiones y el comercio internacional},
+   year = {2000},
+}
+
+@book{HSM2014,
+   author = {C Croarkin and P Tobias},
+   journal = {Retrieved January},
+   title = {NIST/SEMATECH e-handbook of statistical methods},
+   volume = {1},
+   year = {2014},
+}
+
+@Article{tidyverse,
+  title = {Welcome to the {tidyverse}},
+  author = {Hadley Wickham and Mara Averick and Jennifer Bryan and Winston Chang and Lucy D'Agostino McGowan and Romain François and Garrett Grolemund and Alex Hayes and Lionel Henry and Jim Hester and Max Kuhn and Thomas Lin Pedersen and Evan Miller and Stephan Milton Bache and Kirill Müller and Jeroen Ooms and David Robinson and Dana Paige Seidel and Vitalie Spinu and Kohske Takahashi and Davis Vaughan and Claus Wilke and Kara Woo and Hiroaki Yutani},
+  year = {2019},
+  journal = {Journal of Open Source Software},
+  volume = {4},
+  number = {43},
+  pages = {1686},
+  doi = {10.21105/joss.01686},
+  }
+
+@Article{tmap,
+  title = {{tmap}: Thematic Maps in {R}},
+  author = {Martijn Tennekes},
+  journal = {Journal of Statistical Software},
+  year = {2018},
+  volume = {84},
+  number = {6},
+  pages = {1--39},
+  doi = {10.18637/jss.v084.i06},
+  }
+
+@Article{spdep,
+  author = {Roger Bivand},
+  title = {R Packages for Analyzing Spatial Data: A Comparative Case
+    Study with Areal Data},
+  journal = {Geographical Analysis},
+  year = {2022},
+  volume = {54},
+  number = {3},
+  pages = {488-518},
+  doi = {10.1111/gean.12319},
+  }
+
+@Manual{rgeoda,
+  title = {rgeoda: R Library for Spatial Data Analysis},
+  author = {Xun Li and Luc Anselin},
+  year = {2022},
+  note = {R package version 0.0.9},
+  url = {https://CRAN.R-project.org/package=rgeoda},
+  }
+
+@book{Montgomery2012,
+   author = {Montgomery. Douglas C. and Peck. Elizabeth A. and Vining. G. Geoffrey},
+   issn = {03067734},
+   issue = {2},
+   journal = {A JOHN WILEY & SONS, INC., PUBLICATION},
+   title = {Introduction to Linear Regression Analysis, Fifth Edition},
+   volume = {81},
+   year = {2012},
+}
+
+@Article{spatialreg,
+  title = {A Review of Software for Spatial Econometrics in {R}},
+  author = {Roger Bivand and Giovanni Millo and Gianfranco Piras},
+  journal = {Mathematics},
+  year = {2021},
+  volume = {9},
+  number = {11},
+  url = {https://www.mdpi.com/2227-7390/9/11/1276},
+  doi = {10.3390/math9111276},
+  }
+
 @webpage{EF2020,
    author = {Matt},
    journal = {EF Educación Internacional},
