+++ book.bib
@@ -1,5 +1,4 @@
 @book{Fischer2014,
-   abstract = {The Handbook of Regional Science is a multi-volume reference work providing a state-of-the-art knowledge on regional science composed by renowned scientists in the field. The Handbook is intended to serve the academic needs of graduate students, and junior and senior scientists in regional science and related fields, with an interest in studying local and regional socio-economic issues. The multi-volume handbook seeks to cover the field of regional science comprehensively, including areas such as regional housing and labor markets, regional economic growth, innovation and regional economic development, new and evolutionary economic geography, location and interaction, the environment and natural resources, spatial analysis and geo-computation as well as spatial statistics and econometrics.},
    author = {Manfred M. Fischer and Peter Nijkamp},
    doi = {10.1007/978-3-642-23430-9},
    journal = {Handbook of Regional Science},
