+++ README.md
@@ -1,9 +1,11 @@
-Spatial data analysis with R: wrangling, visualization and econometric models is a [complete open course](https://github.com/jaime-pru/Analisis-de-datos-espaciales) in Spanish for the analysis of spatial information in R software. The course is aimed at Spanish-speaking economics undergraduate students interested in developing technical skills for the quantitative analysis required by regional sciences and the spatial approach of economics.
+Spatial Analysis with R: wrangling, visualization and econometric models (Análisis Espacial con R: manejo, visualización y modelos econométricos)is a [full open course](https://github.com/jaime-pru/Analisis-de-datos-espaciales) in Spanish for the analysis of spatial information in R software. The course is intended for Spanish-speaking economics undergraduate students interested in acquiring technical skills for quantitative analysis required by the regional sciences [@Fischer2014] and the spatial approach to economics [@Fujita2000], but it could be useful for other students of social sciences attracted by the topic.  
 
-Its objective is to guide the reader, from a fundamentally practical approach, in the knowledge and handling of techniques for the exploration, analysis and modeling of spatial information through the use of R [@RSoftware] and RStudio, a computer program and programming language focused on statistical analysis. and information visualization and an integrated development environment (IDE), respectively.
+The objective of the book is to guide the student, from a basic practical approach, in the knowledge and handling of spatial information exploration, analysis and modeling techniques through the use of R [@RSoftware] and RStudio, which are, respectively, a program computer science and programming language focused on statistical analysis and information visualization, on the one hand, and an integrated development environment (IDE), on the other.  
 
-The course appears in the form of an electronic book and is structured in five chapters of gradual learning. In chapter 1, the basic elements of R and RStudio are presented through the use of the exploratory data analysis approach, that is, the student is fully introduced to the use of the software to propose and solve questions related to the data structure through various information visualization and manipulation tools.
+The course is presented in the form of an electronic book and it is structured in five chapters of gradual learning. In Chapter 1, the basics of R and RStudio are introduced using the exploratory data analysis approach [@HSM2014] not only with the basic R package, but also with the popular set of tools provided by tidyverse [@ tidyverse], that is, the student is completely introduced to the use of the software to propose and solve questions related to the information structure. Chapter 2 shows how to create various types of choropleth maps and the enormous styling customization flexibility for this purpose via the tmap [@tmap] package. Chapter 3 presents how to carry out an exploratory analysis of spatial data where the student will find how to define the interrelationships that take place in space through the construction of spatial weight matrices and she will learn about spatial autocorrelation and its implications in the information analysis, through various packages such as spdep [@spdep] and rgeoda [@rgeoda]. Meanwhile, in Chapter 4, a very synthetic review of simple regression models is presented, emphasizing the autocorrelation problem that could occur when estimating a linear model with spatial data. Finally, in Chapter 5, two of the different spatial econometric modeling alternatives available in R it is shown, with spatialreg [@spatialreg].  
 
-Chapter 2 shows how to create various types of choropleth maps and the enormous flexibility of style customization that R has for this purpose. Chapter 3 presents the tools to carry out an exploratory analysis of spatial data where the reader will find a way to define the interrelationships that occur in space through the construction of spatial weight matrices and will learn about spatial autocorrelation and its implications in the analysis of information. Meanwhile, in chapter 4, a very synthetic review of simple regression models is presented, emphasizing the problem of autocorrelation that could occur when estimating a linear model with spatial data. Finally, in chapter 5, some of the different spatial econometric modeling alternatives available in R are shown.
[…]
