### 2023-02-01 jaime-pru/Analisis-de-datos-espaciales#1 @jaime-pru
More detailed instructions of how to contribute were given, in English and Spanish. A link for the instructions were placed in the md file.
