### 2023-02-01 27039c8
Detalles sobre contribuciones
