+++ cómo contribuir.md
@@ -1,0 +1,11 @@
+## Cómo contribuir
+
+**Análisis de datos espaciales con R** es un curso de código abierto, dirigido a estudiantes de pregrado en economía y ciencias sociales de habla hispana interesados en en el aprendizaje de herramientas computacionales para el análisis espacial.
+
+Si tienes alguna duda, comentario o encuentras algún error en el [libro electrónico](https://jaime-pru.github.io/Analisis-de-datos-espaciales/), házmelo saber. Aquí hay varias opciones para contribuir con este proyecto, elige la te resulte más cómoda.
+
+a)  Envía un correo a japv@azc.uam.mx indicando dónde identificaste el error o la naturaleza de tu contribución, de la forma más explícita posible.
+
+b) Usa los _Issue_ de GitHub dejando un comentario. Para ello, necesitas crear una cuenta en GitGub para tener acceso a este proyecto y dar click en _New issue_. Asigna un título al comentario, describe la duda, error detectato o tu contribución de la forma más explícita posile y envíalo dando click con _Submit new issue_.  
+
+c) Si cuentas con más conocimientos en GitHub, puedes sugerir los cambios a través de un _Pull Request_ en el repositorio. Para ello, necesistas una cuenta en GitHub y seguir [estas instrucciones](https://www.dataschool.io/how-to-contribute-on-github/) (en inglés).  
