+++ README.md
@@ -1,4 +1,4 @@
-Spatial Analysis with R: wrangling, visualization and econometric models (Análisis Espacial con R: manejo, visualización y modelos econométricos)is a [full open course](https://github.com/jaime-pru/Analisis-de-datos-espaciales) in Spanish for the analysis of spatial information in R software. The course is intended for Spanish-speaking economics undergraduate students interested in acquiring technical skills for quantitative analysis required by the regional sciences [@Fischer2014] and the spatial approach to economics [@Fujita2000], but it could be useful for other students of social sciences attracted by the topic.  
+Spatial Analysis with R: wrangling, visualization and econometric models (Análisis Espacial con R: manejo, visualización y modelos econométricos)is a [full open course](https://jaime-pru.github.io/Analisis-de-datos-espaciales/) in Spanish for the analysis of spatial information in R software. The course is intended for Spanish-speaking economics undergraduate students interested in acquiring technical skills for quantitative analysis required by the regional sciences [@Fischer2014] and the spatial approach to economics [@Fujita2000], but it could be useful for other students of social sciences attracted by the topic.  
 
 The objective of the book is to guide the student, from a basic practical approach, in the knowledge and handling of spatial information exploration, analysis and modeling techniques through the use of R [@RSoftware] and RStudio, which are, respectively, a program computer science and programming language focused on statistical analysis and information visualization, on the one hand, and an integrated development environment (IDE), on the other.  
 
