### 2023-02-01 jaime-pru/Analisis-de-datos-espaciales#3 @jaime-pru
Se atendieron los siguientes elementos:

- Se añadió el título como parte del cuerpo del texto.
- Se brindó una definición muy breve de ciencias regionales y se remite a la estudiante a un material en castellano al respecto.
- Se colocó una referencia específica, aunque breve, sobre datos espaciales en relación con las ciencias regionales, aunque no en el párrafo segundo, sino en el cuarto.
- Se dio el título de la obra _Handbook..._ en castellano y se ofrece a la estudiante un material en el idioma.
- Se eliminó la mención sobre la universidad en México, pues se consideró más oportuno brindar esa información en los datos del autor.
- Se eliminó el enlace citado para evitar confusiones y la referencia al curso pues, en efecto, no era pertinente.
- Se dejó sólo el apellido de los autores citados.
- Se indicó, en forma breve, que R es software libre.
- Los nombres de los capítulos fueron colocados en mayúsculas y se colocó el vínculo respectivo.
- Se mencionó en el prólogo no sólo la licencia, sino que se incluyó el enlace al archivo md que abunda sobre cómo contribuir con el proyecto.
- Se trató de usar un lenguaje no sexista. Se optó por uniformizar en "la estudiante".
