+++ Contributing.md
@@ -1,11 +1,0 @@
-## How to Contribute
-
-**Análisis de datos espaciales con R** is an open-source course, aimed to spanish-speakers last year undergraduate economics and social sciences students interested in learning computational tools to spatial analysis.
-
-If you have any question, comment or you find an error in the [e-boook](https://jaime-pru.github.io/Analisis-de-datos-espaciales/), please let me know it. Here, you have some options to contribute to this project, choose whatever you feel more comfortable.
-
-a) Mail me to japv@azc.uam.mx and let me know as clear as possible the identified error or how you want to contribute to the project.
-
-b) Use _Issue_ options on GitHub and leave a comment. To do so, you need a GitHub account to access the repo project and select _New issue_. Give it a title, desscribe dubts, errors o suggestions as clear as possible and submit clicking on _Submit new issue_.  
-
-c) If you are comfortable enough using GitHub, you may suggest changes trough a _Pull Request_ in the repo. As in b, you need a GitHub account and follow this [instructions](https://www.dataschool.io/how-to-contribute-on-github/) (in Englis).
