+++ book.bib
@@ -194,8 +194,16 @@ For this notebook, we will continue to use the socioeconomic data for 55 New Yor
    year = {2015},
 }
 
+@book{Correa2020,
+   author = {Ronny Correa-Quezada and Mario Polèse and Fernando Rubiera Morollón},
+   publisher = {Alpha Editorial},
+   title = {Economía urbana y regional: Teorías y evidencias sobre el desarrollo territorial},
+   year = {2020},
+}
+
+
 @book{Asuad2001,
-   author = {Normand E. Asuad Sanén},
+   author = {Normand Asuad},
    city = {Puebla},
    edition = {1},
    journal = {Science},
