### 2023-02-01 jaime-pru/Analisis-de-datos-espaciales#4 @jaime-pru
Se atendieron las siguientes observaciones al Capítulo 1:

- Se añadió la frase "lo primero será descargar e instalar R y luego descargar e instalar RStudio" para aclarar el proceso de instalación.
- Se eliminó en todos los casos el espacio en blanco en "RStudio".
- Se cambió a "Figura", con mayúsculas.
- Se cambió la Figura 1.1. por el modo claro.
- Se cambió “Hello world” simplemente por “Hola"
- Se indicó que, además de copiar, hay que pegar el código en la consola.
- Sobre el uso del símbolo  #, se añadió un comentario sobre lo que significa en la primera función en que se utiliza.
- Para obtener ayuda, se corrigió indicando ??tmap.
- Además de la referencia en inglés de _R for Data Science_, se brinda la versión en castellano.
-  Se incluyó la liga a las versiones en español de las hojas de trucos.
- Respecto a la notación paquete::función, se añadió un comentario en la función que por fines didácticos se utiliza de esta forma,  pero que no es necesario.
- Se indica cómo dar crédito a los autores de los paquetes con citation(“tmap”)
- Se corrigieron los errores ortotipográficos señalados.
