+++ 01-intro.Rmd
@@ -70,7 +70,7 @@ Para crear objetos, en R se utiliza el signo "menor que" seguido de un guion par
 ```{r}
 #Nota: el símbolo de # representa un comentario y no se ejecuta con el código
 x <- 3 #Numérico
-y <- "Hola Mundo" #Carácter 
+y <- "Hola" #Carácter 
 z <- FALSE # Lógico
 ```
 
@@ -179,6 +179,16 @@ Se dijo que R es un programa especializado en el análisis estadístico y la rep
 -   `tmap`: ofrece un enfoque flexible, basado en capas y fácil de usar para crear mapas temáticos.\
 -   `spatialreg`: paquete para la elaboración de regresiones con componentes espaciales.
 
+Los paquetes, que expanden las capaciades de la versión base de R, son elaborados por múltiples colaboradores. Para conocer quines están detrás de ello, podemos usar la función `citation()`, que nos muestra cómo dar crédito a los autores de los paquetes cuando los usamos.
+
+```{r include=FALSE}
+library(tmap)
+```
+
+```{r}
+citation("tmap")
+```
+
 Para poder hacer uso de los paquetes que amplían el potencial de R es necesario descargarlos, instalarlos y, en cada sesión de trabajo, llamarlos. Para la descarga e instalación podemos ir a Archivo (*File*) en la cinta de menú que se localiza en la parte superior de tu entorno de trabajo y seleccionar *tools/install package*. A continuación, deberás colocar el nombre del paquete deseado y dar click en *instalar*. O bien, alternativamente, puedes ir a la sección de Archivo, en la zona inferior derecha, seleccionar la pestaña *Packages* y a continuación el ícono de *Install*. Una tercera manera de instalar paquetes es desde la consola (sección inferior izquierda), para instalar un paquete a la vez:
 
 ```{r, eval=FALSE}
@@ -219,7 +229,7 @@ help(matrix)
 help(mean)
 ```
 
-En la cinta de menú, en la sección ayuda, *Help*, encontrarás una serie de materiales muy útiles para familiarizarse con los paquetes instalados. Estos materiales reciben el nombre de "hojas de trucos", *Cheatsheet*. Se recomienda ampliamente revisar cada uno de ellos. Además, R es un programa con un sin número de entusiastas usuarios y con un amplio soporte técnico por lo que cuando te encuentres con una dificultad para usar algún paquete o función puedes remitite al sitio de [Ayuda del R](https://www.r-project.org/help.html), o bien, a alguno de los repositorios especializados para presentar y resolver dudas como [Stackoverflow](https://es.stackoverflow.com/). Para ilustrar esto, ingresa al sitio de Stackoverflow y coloca en la búsqueda "histograma en R".
+En la cinta de menú, en la sección ayuda, *Help*, encontrarás una serie de materiales muy útiles para familiarizarse con los paquetes instalados. Estos materiales reciben el nombre de "hojas de trucos", *Cheatsheet*. Se recomienda ampliamente revisar cada una de ellas, incluidas las [versiones en epañol](https://posit.co/resources/cheatsheets/?type=translations/). Además, R es un programa con un sin número de entusiastas usuarios y con un amplio soporte técnico por lo que cuando te encuentres con una dificultad para usar algún paquete o función puedes remitite al sitio de [Ayuda del R](https://www.r-project.org/help.html), o bien, a alguno de los repositorios especializados para presentar y resolver dudas como [Stackoverflow](https://es.stackoverflow.com/). Para ilustrar esto, ingresa al sitio de Stackoverflow y coloca en la búsqueda "histograma en R".
 
 ## Cargar bases de datos
 <div style="text-align: justify">
[…]
