### 2023-02-01 jaime-pru/Analisis-de-datos-espaciales#5 @jaime-pru
Correcciones hechas al Capítulo 2:

- Se prefirió cambiar la referencia al material licenciado y, en su lugar, se proporcionó un vínculo a otro material.
- Se incluyó de forma amplia las referencias al material de Víctor Olaya.
- Sobre el término “georreferenciada”, se trató de dar una definición que, aunque sucinta, recoge los elementos relevantes para el contenido que es expuesto. Además, se proporcionaron un par de ligas a materiales para abundar sobre el problema de la georreferencia.
- Para mejor definir los datos vectoriales y ráster, se toman algunos elementos de la definición de Olaya para tratar de resolver el error. Se remite a la estudiante al material recomendado.
- Sobre los datos vectoriales, se trató de mejorar la redacción para evitar la posible confusión entre el uso de las primitivas de líneas, puntos o polígonos.
- Sobre los datos raster, se toman algunos elementos de la definición de Olaya para tratar de resolver el error. Se remite a la estudiante al material recomendado.
- Sobre el desarrollo de ESRI de los SHP, se modificó la redacción de este párrafo para evitar la imprecisión señalada.
- Sobre el archivo .prj, En la nota al pie número 2 de este capítulo se señala los otros archivos, incluidos el PRJ.
- En la exposición del paquete sf, se simplificó su exposición y uso.
- En la definición de mapa coroplético, en efecto, había un error. Se buscó aclarar los términos y se remitió a la estudiante al libro recomendado.
- Sobre el uso de las paletas de colores, se amplió la información sobre tres tipos de paletas y se abundó sobre su posible uso. Se añadió también la referencia recomendada.
- Para que la leyenda de los mapas quede en castellano, se incluyó una breve explicación de cómo modificar dicho elemento.
- Se corrigió la redacción para usar un lenguaje no sexista.
- Se buscó corregir todos los errores tipográficos identificados.
