### 2023-02-01 jaime-pru/Analisis-de-datos-espaciales#6 @jaime-pru
Correcciones del Capítulo 3:

- Sobre la introducción del capítulo, se escribió una pequeñísima introducción para ilustrar intuitivamente mediante un mapa, la idea de la autocorrelación espacial. Además, se eliminó la referencia a la contribución de Snow por considerar que, en efecto, podría generar confusión.
- Sobre el concepto de autocorrelación, se trató de explicar intuitivamente su sentido y definición.
- Se eliminó la maña exposición sobre el paquete RGDAL, por considerarla accesoria.
- No se logró identificar, en la sección 3.4, el formato erróneo a un texto normal.
- Se indicó el significado de LISA y su sentido en castellano.
- En la figura de Sección 3.5, se añadió la clasificación de cuadrantes en la leyenda.
- Se corrigieron los errores tipográficos.
