### 2023-02-01 jaime-pru/Analisis-de-datos-espaciales#7 @jaime-pru
Correcciones al Capítulo 4:

- Sobre los materiales licenciados, se cambió a otra referencia se proporcionó vínculo al material.
- Sobre el sentido no espacial de este capítulo, se incluyó una pequeña introducción al capítulo donde se señala el alcance de éste.
- Se corrigió la referencia incorrecta en el archivo .bib: C., A., y Geoffrey 2012.
- Se incluyó el año en la referencia a Dalpiaz.
- Se incluyó una mención sobre en qué consiste un modelo econométrico y las diferencias con otros modelos (teóricos y matemáticos).
- Se incluye el gráfico de dispersión sugerido.
- Se corrigieron los errores tipográficos.
