### 2023-02-01 jaime-pru/Analisis-de-datos-espaciales#8 @jaime-pru
Correcciones al Capítulo 5:

Este capítulo se modificó de forma considerable, tratando de atender a los comentarios de las revisoras. Gracias a sus críticas, me parece el capítulo ganó claridad y precisión.

- Se escribió una explicación más detallada y más didáctica sobre el llamado modelo general.
- Del multiplicador de Lagrange, se trató de explicar, sin llegar a su formalización, en qué consiste el estadístico, usando algunas comparaciones.
- Del modelo es “anidado”, se trató de aclarar mejor el modelo y explicarlo con detalle.
- Se reeditaron las figuras 1, 2 y 3, con una mejor resolución.
- Se corrigió la cita de Anselin y Rey, no se colocó el link después de la primera mención.
- Sobre las versiones robustas del multiplicador de Lagrange, se trató de hacer una exposición más cuidadosa de la ruta de interpretación de los modelos con el uso de tales estadísticos.
- Sobre la decisión de los modelos, se hizo un esfuerzo por lograr una exposición más cuidadosa y guiada, se incluyó de forma explícita la secuencia que la estudiante debe seguir y se brindaron más ejemplos.
- Sobre los modelos econométricos, se trató de brindar una mejor exposición del tipo de modelos econométricos usados y se puntualizó la terminología en la sección aclarando que todo lo expuesto, en el capítulo 4 y 5, son modelos econométricos.
- Antes del final, se hace una recapitulación donde se usan herramientas de los capítulos previos, tratando de unir todos los elementos mínimos necesarios para una adecuada secuencia de interpretación.
- Se corrigieron los errores orto-tipográficos.
