+++ book.bib
@@ -139,7 +139,7 @@ For this notebook, we will continue to use the socioeconomic data for 55 New Yor
 }
 
 @book{Montgomery2012,
-   author = {Montgomery. Douglas C. and Peck. Elizabeth A. and Vining. G. Geoffrey},
+   author = {Douglas C. Montgomery and Elizabeth A. Peck. and G. Geoffrey Vining.},
    issn = {03067734},
    issue = {2},
    journal = {A JOHN WILEY & SONS, INC., PUBLICATION},
@@ -148,11 +148,14 @@ For this notebook, we will continue to use the socioeconomic data for 55 New Yor
    year = {2012},
 }
 
+
 @webpage{Dalpiaz2022,
    author = {David Dalpiaz},
    title = {Applied Statistics with R},
    url = {http://daviddalpiaz.github.io/appliedstats/},
+   year = {2022},
 }
+
 
 @book{Gujarati2011,
    author = {Damodar Gujarati and N Dawn C. Porter},
