### 2023-02-01 jaime-pru/Analisis-de-datos-espaciales#9 @jaime-pru
En atención a los comentarios de la revisora, se enlistan los cambios hechos:

- Se modificó “apuntamos” a “señalamos”.
- Se indicó, con relación a los datos raster, que su fuente es diversa.
- Se señaló que el atributo tiempo puede estar ausente de algunos datos.
- Se cambió la línea indicada señalando el tipo de información espacial usada.
- Se corrigió el error tipográfico de la línea 92
- Se ofreció una exposición más cuidada sobre los SRC y se brindó información sobre cómo es posible cambiarlos, aunque no se abundó en ello.
- Se indicó, con un ejemplo, para qué es útil el cálculo de centroides.
- Sobre los mapas de percentiles, se indicó su utilidad en la identificación de valores atípicos.
