### 2023-02-01 jaime-pru/Analisis-de-datos-espaciales#10 @jaime-pru
Estos son los cambios y correcciones que corresponden a las observaciones de la revisora respecto a la introducción:

- Se expuso con más cuidado qué son los factores y para qué sirven.
- Se comentó la distinción entre el paquete `base` y los paquetes que expanden las posibilidades de R.
- Se brindan unas líneas introductorias donde se apunta el alcance del capítulo.
