### 2023-02-01 jaime-pru/Analisis-de-datos-espaciales#11 @jaime-pru
Respecto a los comentarios sobre el capítulo 4, se corrigieron los siguientes elementos:

- Se corrigió la redacción sobre las preguntas causales.
- Se trató de clarificar la exposición sobre causalidad y se incorporó el ejemplo sugerido.
- Se eliminaron los tres puntos que estaban incorrectamente ubicados.
- Se indicó, a través de una exposición más clara, el sentido de los modelos lineales simples.
- Se incluyó una pregunta con el tópico recomendado sobre la linealidad en los parámetros.
- Sobre los supuestos del modelo de regresión lineal, se remite a la estudiante a un material clásico donde puede profundizar en su estudio, en la medida en que lo que se busca con este capítulo es que ella tenga presente el sentido del supuesto de independencia en los términos de error.
- Se abundó un poco más sobre las pruebas de hipótesis, tratando de dar un ejemplo e incorporando la distinción solicitada entre hipótesis nula y alternativa.
- Se trató de hacer una síntesis de ideas al final del capítulo en relación con el análisis causal y el alcance del análisis de regresión para subsanar la debilidad identificada por la revisora.
