### 2023-02-01 jaime-pru/Analisis-de-datos-espaciales#12 @jaime-pru
Este capítulo fue reestructurado de forma considerable, gracias a las observaciones de las revisoras, por lo que se añadieron secciones, se ampliaron otras y en general se modificó la exposición.

Se recordó en una introducción añadida qué es lo que se pretende con el análisis de regresión y cómo se busca, a través de la econometría espacial, atender la violación a uno de los supuestos del modelo lineal.

Se definió con cuidado el concepto de econometría espacial y se expuso cada uno de los efectos espaciales que integra el modelo general.  Además, se desglosó con más detalle el proceso para la estimación de los modelos espaciales, siguiendo las recomendaciones dadas, dividiendo por pasos. Por otro lado, se abundó sobre la interpretación de los dos modelos desarrollados.

Finalmente, se añadió una sección final donde se busca hacer una recapitulación de los contenidos tratados en el libro y se enlistan recomendaciones de lectura adicional.
