+++ book.bib
@@ -35,10 +35,26 @@
    year = {2007},
 }
 
+@webpage{ESRI2023,
+   author = {ESRI},
+   journal = {ArcGIS for Desktop-Proyecciones cartográficas.},
+   month = {2},
+   title = {Sistema de coordenadas geográfica},
+   url = {https://desktop.arcgis.com/es/arcmap/10.3/guide-books/map-projections/geographic-coordinate-system.htm},
+   year = {2023},
+}
+
+@web_page{ESRI2023b,
+   author = {ESRI},
+   journal = {ArcGIS for Desktop-Sistemas de coordenadas proyectadas.},
+   month = {2},
+   title = {Qué son los sistemas de coordenadas proyectadas},
+   url = {https://desktop.arcgis.com/es/arcmap/10.3/guide-books/map-projections/about-projected-coordinate-systems.htm},
+   year = {2023},
+}
 
 
 @article{Chasco2003,
-   abstract = {El objetivo fundamental de esta tesis consiste en proponer la predicción\nde datos espaciales como parte de la econometría espacial, a partir de\nuna metdología fundamentada en instrumentos exploratorios y confirmatorios\npropios de esta disciplina, En concreto, se desarrolla una subdisciplina\nde la predicción espacial, denominada predicción-extrapolación espacial,\ndistinta del método geoestadístico de interpolación espacial, entendida\ncomo el conjunto de métodos inferenciales destinados a obtener datos microterritoriales\na partir de información espacial agregada.\n\nLa consecución de este objetivo general exige una presentación y sistematización\ndel conjunto de técnicas de análisis exploratorio y confirmatorio de la\neconometría espacial, realizada en el cuerpo central de esta obra, con\nla ayuda de un capítulo completo y actualizando de Bibliografía, así como\nde unos Anexos, que amplían algunas ideas sin romper el hilo de la exposición.\n\nLa metodología propuesta se evalúa a través de un ejercicio de predicción-extrapolación\nde la renta familiar disponible de los municipios de la Comundiad de Madrid,\nen el que se hacen confluir tanto la utilización de técnicas de econometría\nespacial (a través de un "software" adecuado), como la experiencia de la\nautora en análisis microterritoriales, como Directora A. Del Área de Economía\nEspacial Microterritorial del Instituto Lawrence R. Klein (Universidad\nAutónoma de Madrid), co-dirigido por D. José Vicéns Otero, director de\nesta tesis.\n\nDado el interés y el cierto desconocimiento que aún pesa sobre el tema\nde la econometría espacial, esta tesis se plantea también, como objetivo,\nser de utilidad para todos los investigadores sociales interesados en el\nanálisis territorial, en cualquiera de sus formas, así como contribuir,\nen la medida de los posible, a la difusión de las técnicas de econometría\nespacial en nuestro país, sobre todo en el ámbito de los estudios},
    author = {Coro Yrigoyen Chasco},
    issue = {April 2003},
    journal = {Consejería de Economía e Innovación Tecnológica},
@@ -47,7 +63,6 @@
 }
 
 @book{Lovelace2019,
[…]
