+++ 01-intro.Rmd
@@ -1,5 +1,8 @@
 # R: Una introducción desde la exploración de información
 
+En este capítulo buscamos cubrir dos objetivos. El primero corresponde a que la estudiante se familiarice con los diversos elementos que componen la intefaz de trabajo en RStudio y conozca la definición de objeto y sus tipos, en el contexto de los lenguajes de programación. El segundo es que conozca cómo desarrollar un análisis exploratorio de datos, a través de un tipo de objetos específicos (los `dataframe`).
+
+
 ## R: Lenguaje de programación y plataforma de análisis
 <div style="text-align: justify">
 R es un lenguaje de programación orientado a objetos, libre y de código abierto y además un ambiente enfocado al análisis estadístico y gráfico. Que sea libre significa, a diferencia de otros programas, que cualquier persona puede usarlo, redistribuirlo o modificarlo, sin necesidad de contar con una licencia pago.  
@@ -157,14 +160,12 @@ rbind(x,y)
 
 #### Factores
 
-Son un tipo especial de vectores y son utilizados para representar información categórica, lo que permite organizarla en niveles para analizarla mejor:
+Son un tipo especial de vectores y son utilizados para representar información categórica, lo que permite organizarla en niveles para analizarla mejor. Para mejor entender esto, basta recordar la definición de variable. Las “variables” o “atributos” son medidas o características de los elementos que conforman la población. Las medidas son cuantitativas y las características son cuantitativas. Justamente los factores se refieren a variables que recogen cualidades. Por ejemplo, si la variable responde a la pregunta "¿el paciente recibió la vacuna contra COVID-19?" y hay dos posibles respuestas, es recomendable que en lugar de designar cada alernativa como 1 o 2  se usen variables categóricas como "Vacuna aplicada" y "Vacuna no aplicada", es decir, es recomendable almacenar la información como factor:
 
 ```{r}
-x <- factor(c("yes","no","no","yes","no"))
+x <- factor(c("Vacuna no aplicada","Vacuna no aplicada","Vacuna no aplicada","Vacuna no aplicada","Vacuna no aplicada"))
 x
 ```
-
-Por ejemplo, si la variable responde a la pregunta ¿el paciente recibió la vacuna contra COVID-19? y hay dos posibles respuestas, es recomendable que en lugar de designar cada alernativa como 1 o 2  se usen variables categóricas como "Vacuna aplicada" y "Vacuna no aplicada", es decir, es recomendable almacenar la información como factor.
 
 #### _Data frame_ o arreglo de datos
 
@@ -174,7 +175,9 @@ Cuando arribemos un poco más adelante al proceso de importación de informació
 
 ### Paquetes
 
-Se dijo que R es un programa especializado en el análisis estadístico y la representación gráfica, pero R no sólo se limita a lo que ofrece cuando lo descarga por primera vez. Uno de los elementos que hace de R una potente herramienta es la posibilidad de ampliar su potencial a través de la instalación de paquetes que expanden sus funciones básicas. Existen paquetes de R para múltiples campos disciplinares y especialidades, por ejemplo estas notas se sirven de los paquetes especializados en el análisis espacial, como se observará en los capítulos subsecuentes. De momento mencionemos sólo un par:
+Se dijo que R es un programa especializado en el análisis estadístico y la representación gráfica, pero R no sólo se limita a lo que ofrece cuando lo descarga por primera vez, es decir a su paquete `base` o paquete base. Por ejemplo, dentro del paquete `base` se cuenta con la función para calcular la media de una muestra, `mean()`, entre otras tantas.
+
[…]
