+++ book.bib
@@ -44,7 +44,7 @@
    year = {2023},
 }
 
-@web_page{ESRI2023b,
+@webpage{ESRI2023b,
    author = {ESRI},
    journal = {ArcGIS for Desktop-Sistemas de coordenadas proyectadas.},
    month = {2},
