+++ packages.bib
@@ -11,7 +11,7 @@
   title = {bookdown: Authoring Books and Technical Documents with R Markdown},
   author = {Yihui Xie},
   year = {2022},
-  note = {R package version 0.31},
+  note = {R package version 0.30},
   url = {https://CRAN.R-project.org/package=bookdown},
 }
 
@@ -19,7 +19,7 @@
   title = {knitr: A General-Purpose Package for Dynamic Report Generation in R},
   author = {Yihui Xie},
   year = {2022},
-  note = {R package version 1.41},
+  note = {R package version 1.40},
   url = {https://yihui.org/knitr/},
 }
 
@@ -27,7 +27,7 @@
   title = {rmarkdown: Dynamic Documents for R},
   author = {JJ Allaire and Yihui Xie and Jonathan McPherson and Javier Luraschi and Kevin Ushey and Aron Atkins and Hadley Wickham and Joe Cheng and Winston Chang and Richard Iannone},
   year = {2022},
-  note = {R package version 2.19},
+  note = {R package version 2.18},
   url = {https://CRAN.R-project.org/package=rmarkdown},
 }
 
