+++ _bookdown.yml
@@ -8,5 +8,5 @@ language:
     chapter_name: ""
 output_dir: docs
 rmd_files: ["index.Rmd","01-intro.Rmd","02-Mapas.Rmd","03-analisis-espacial-I-rgeoda.Rmd","04-Regresion-lineal.Rmd","05-Analisis-espacial-II.Rmd","07-sobre-el-autor-y-creditos.Rmd","08-references.Rmd"]
-csl: css/harvard-cite-them-right-11th-edition.csl
+csl: css/regional-science-and-urban-economics.csl
   
