+++ index.Rmd
@@ -28,7 +28,7 @@ Análisis de datos espaciales con R {-}
 <div style="text-align: justify">
 <img src="Portada.png" width="300" height="425" alt="Portada del libro electrónico 'Análisis de datos espaciales con R'" align="right" style="margin: 0 1em 0 1em" /></a>
   
-# Prólogo
+# Prólogo {-}
 
 Las ciencias regionales, un conjunto amplio de especialidades que estudian los fenómenos sociales, económicos y políticos que tienen una dimensión espacial[^1], no son ya una disciplina menor, pasaron de ser una suerte de espacio para la experimentación de teorías y métodos que no tenían cabida en las corrientes principales de los cuerpos científicos plenamente reconocidos y se han ido abriendo paso y consolidando como una disciplina totalmente reconocida. Las ciencias regionales pueden ser entendidas como un conjunto de campos disciplinares en los que el punto en común es el interés dado al espacio (ya sea este entendido como regiones, ciudades, infraestructuras, medio ambiente) en los fenómenos sociales tales como las disparidades socioeconómicas, el crecimiento urbano y regional o el transporte y la logística [@Nijkamp2015]. Las ciencias regionales han ido adquiriendo “una amplia orientación multidisciplinaria sobre temas regionales y urbanos, combinando y siendo un complemento a la economía regional, geografía social y económica, economía urbana, ciencia del transporte, ciencia ambiental, ciencia política y teoría de la planificación” [@Fischer2014]. La realidad es que las ciencias regionales han evolucionado como un campo de investigación en pleno derecho. Tras un largo recorrido que incluso se podría remontar a los inicios de la economía como ciencia, comenzando con los trabajos de Adam Smith que trataban algunos elementos relacionados con la localización de actividades comerciales, pasando por los estudios clásicos de Von Thünen, Weber, Lösch (todos ellos relacionados con la denominada teoría de la localización), hasta llegar a la consolidación de las ciencias regionales como disciplina gracias al impulso de [Isaard](https://www.narsc.org/newsite/internal/isard/) que trata de brindar una perspectiva teórica y metodológica coherente con un fuerte soporte empírico desde una perspectiva multidisciplinaria. Las contribuciones de W. Isaard abarcan una multitud de campos tales como la ecología, el estudio de los transportes e incluso el manejo de conflictos sociales.
 
