### 2023-02-02 fa3aeb9
borre la imagen, por eso no aparecia
