+++ docs/reference-keys.txt
@@ -1,3 +1,4 @@
+fig:unnamed-chunk-2
 fig:unnamed-chunk-32
 fig:unnamed-chunk-113
 fig:unnamed-chunk-114
