### 2023-02-02 30598bb
Act. logo de licencia
