+++ index.Rmd
@@ -6,6 +6,7 @@ knit: "bookdown::render_book"
 site: bookdown::bookdown_site
 output:
  bookdown::bs4_book:
+  lang: es
   theme:
     primary: "#3860b6"
     base_font: "Segoe UI"
