+++ 01-intro.Rmd
@@ -23,7 +23,7 @@ En tanto, RStudio puede ser descargado de [aquí](https://www.rstudio.com/). Ve 
 
 Al abrir RStudio se mostrará una pantalla como la que aparece en la Figura 1.1. La pantalla principal de RStudio está dividida en cuatro partes. En caso de que el recuadro superior izquierdo no aparezca, da clic en el ícono de la hoja en blanco con el signo de más (*New file*, parte superior izquierda de la pantalla) para que se despliegue.
 
-```{r, echo=FALSE, fig.align='center',out.width='100%',fig.cap='Panel principal', fig.alt = "Texto alternativo"} 
+```{r, echo=FALSE, fig.align='center',out.width='100%',fig.cap='Panel principal', fig.alt = ""El diseño de la pantalla principal de RStudio está organizado en cuatro secciones distintas, cada una con un propósito específico. Fuente u origen, Ambiente\Historial, Consola y Archivo\Paquetes\Ayuda"} 
 knitr::include_graphics("recursos intro/Figura_1-Entorno.jpg")
 
 ```
@@ -326,9 +326,9 @@ covid_zmvm <- utils::read.csv(ruta)
 
 ### Cargar bases con la funcionalidad importar
 
-Otra manera de cargar archivos de Excel o de texto (formato CSV) es ubicarnos en la sección de ambiente, ventana superior derecha, y seleccionar el ícono *Import Dataset*. Se desplegará un menú donde habremos de elegir el tipo de archivo que se desee importar. R no sólo permite importar archivos de texto o Excel, también bases de SPSS o STATA. Siguiendo con la ilustración relativa a libros de Excel, en la ventana que se despliega, habremos de señalar la ruta exacta o directorio donde está nuestro archivo, incluyendo el nombre y extensión de éste. Al pulsar en el botón actulizar (*Update*) se desplegara una vista previa de la base. Si aparece de forma correcta, seleccionaremos importar (*Import*), en caso contrario, se deberá modificar las opciones de importación. Después de seleccionar Import Dataset/From Excel y elegir la ubicación del archivo que deseas importar, deberías ver en tu pantalla una imagen como la de la Figura 1.3:
+Otra manera de cargar archivos de Excel o de texto (formato CSV) es ubicarnos en la sección de ambiente, ventana superior derecha, y seleccionar el ícono *Import Dataset*. Se desplegará un menú donde habremos de elegir el tipo de archivo que se desee importar. R no sólo permite importar archivos de texto o Excel, también bases de SPSS o STATA. Siguiendo con la ilustración relativa a libros de Excel, en la ventana que se despliega, habremos de señalar la ruta exacta o directorio donde está nuestro archivo, incluyendo el nombre y extensión de éste. Al pulsar en el botón actulizar (*Update*) se desplegara una vista previa de la base. Si aparece de forma correcta, seleccionaremos importar (*Import*), en caso contrario, se deberá modificar las opciones de importación. Después de seleccionar Import Dataset/From Excel y elegir la ubicación del archivo que deseas importar, deberías ver en tu pantalla una imagen como la de la Figura 1.2:
 
-```{r, echo=FALSE, fig.align='center',out.width='120%', fig.cap='Importar datos de excel'} 
+```{r, echo=FALSE, fig.align='center',out.width='120%', fig.cap='Importar datos de excel', fig.alt="Ventana panel que muestra una vista previa de los datos que se estan importando, incluye además, la opción del nombre que desea importar, la hoja que debe importar, el rango de la tabla, etc."} 
[…]
