+++ 01-intro.Rmd
@@ -329,7 +329,7 @@ covid_zmvm <- utils::read.csv(ruta)
 
 Otra manera de cargar archivos de Excel o de texto (formato CSV) es ubicarnos en la sección de ambiente, ventana superior derecha, y seleccionar el ícono *Import Dataset*. Se desplegará un menú donde habremos de elegir el tipo de archivo que se desee importar. R no sólo permite importar archivos de texto o Excel, también bases de SPSS o STATA. Siguiendo con la ilustración relativa a libros de Excel, en la ventana que se despliega, habremos de señalar la ruta exacta o directorio donde está nuestro archivo, incluyendo el nombre y extensión de éste. Al pulsar en el botón actulizar (*Update*) se desplegara una vista previa de la base. Si aparece de forma correcta, seleccionaremos importar (*Import*), en caso contrario, se deberá modificar las opciones de importación. Después de seleccionar Import Dataset/From Excel y elegir la ubicación del archivo que deseas importar, deberías ver en tu pantalla una imagen como la de la Figura 1.2:
 
-```{r, echo=FALSE, fig.align='center',out.width='120%', fig.cap='Importar datos de excel', fig.alt="Ventana panel que muestra una vista previa de los datos que se estan importando, incluye además, la opción del nombre que desea importar, la hoja que debe importar, el rango de la tabla, etc."} 
+```{r, echo=FALSE, fig.align='center',out.width='120%', fig.cap='Importar datos de excel', fig.alt='Ventana panel que muestra una vista previa de los datos que se estan importando, incluye además, la opción del nombre que desea importar, la hoja que debe importar, el rango de la tabla, etc.'} 
 knitr::include_graphics("recursos intro/Figura_3-Importar.jpg")
 ```
 
@@ -390,7 +390,7 @@ La construcción de gráficas con `ggplot2` se hace a partir de una suerte de "c
 
 Así pues, se requieren tres elementos para la construcción de una gráfica con `ggplot2`: datos, geometría y sistema de coordenadas. Dentro de la geometría se especifican no sólo la o las variables a representar, también algunos elementos estéticos como colores. En nuestro caso, la información está contenida en la base `covid_zmvm` pero habrá que especificar qué tipo de gráfica deseamos y sus elementos estéticos. Veamos esto en acción:
 
-```{r message=FALSE, warning=FALSE, fig.alt= "Gráfico de dispersión que representa la relación entre los casos positivos (pos_hab) en el eje x y el número promedio de grados cursados (grad) en y. El gráfico muestra la mayor agrupación de dato por debajo de 10 en pos_hab y 12 en grad" }
+```{r message=FALSE, warning=FALSE, fig.alt= 'Gráfico de dispersión que representa la relación entre los casos positivos (pos_hab) en el eje x y el número promedio de grados cursados (grad) en y. El gráfico muestra la mayor agrupación de datos por debajo de 10 en pos_hab y 12 en grad' }
 library(tidyverse)
 ggplot2::ggplot(data=covid_zmvm)+
   ggplot2::geom_point(aes(x=grad,y=pos_hab))
@@ -413,7 +413,7 @@ Por favor, revisa la hoja de trucos del paquete para que te familiarices con las
 
 Añadamos una "capa" adicional a nuestra gráfica, combinando otra geometría como una línea de ajuste (pongamos atención en el uso del signo de más):
 
[…]
