+++ index.Rmd
@@ -27,9 +27,9 @@ knitr::opts_chunk$set(echo = TRUE)
 Análisis de datos espaciales con R {-}
 =======
 
-<div style="text-align: justify">
-<img src="Portada.png" width="300" height="425" alt="Portada del libro electrónico 'Análisis de datos espaciales con R'" align="right" style="margin: 0 1em 0 1em" /></a>
 
+<img src="Portada.png" width="300" height="425" alt="Portada del libro electrónico 'Análisis de datos espaciales con R'" align="right" style="margin: 0 1em 0 1em" /></a>
+<div style="text-align: justify">
 <a rel="license" href="https://creativecommons.org/licenses/by/4.0/deed.es">
   <img alt="Licencia Creative Commons" style="border-width:0" src="https://i.creativecommons.org/l/by-sa/4.0/88x31.png"/>
 </a>
