### 2023-02-08 jaime-pru/Analisis-de-datos-espaciales#2 @jaime-pru
En términos de accesibilidad, se tomaron en consideración los comentarios hechos, por lo que se añadió texto alternativo a las imágenes, además de que se usó lenguaje no sexista.

Se corrigió lo relativo a la licencia.
