### 2023-05-20 #173 @labarba
hi @jaime-pru — thanks for your patience. Your handling editor has moved on from academia to industry, and has stepped down from the board. She did say that she would finish this submission, so I hope she'll pop in to check your revisions.

A heads-up on the final steps that will be needed once she recommends acceptance—you will be asked to do the following:

- [ ] Make a tagged release in your repository, and list the version tag here.
- [ ] Archive the release in Zenodo (or a similar service)
