### 2023-05-22 #173 @jaime-pru
Hi, @labarba, thank you very much. I appreciate @allisonhorst's time and disposition to complete the processes. Meanwhile, I will work on the listed tasks.

Thanks again to @nmorandeira and @hernandezcortes for their comments and observations.
