### 2023-05-29 #173 @allisonhorst
@labarba can you advise on the above re: PDF failing to compile - the authors have included both paper-en.md and paper-es.md as English and Spanish versions in the repo, which I assume is creating the issue. Should they update one to just paper.md?
