### 2023-05-29 #173 @allisonhorst
Hi all, thank you for your patience! Huge thanks to @nmorandeira and @hernandezcortes for such thoughtful reviews and feedback, and to @jaime-pru for the thorough revision and summary of changes made.

I am excited to recommend acceptance to JOSE.

@jaime-pru can you please let me know (or check off above) when the tasks above to make the tagged release & Zenodo DOI are complete? Thank you!
