### 2023-05-30 #173 @jaime-pru
Hi, @allisonhorst. I believe that the tasks listed by @labarba have now been completed:

1. Make a tagged release in your repository, and list the version tag here.
The tagged release is v.1.0.0 and can be checked [here](https://github.com/jaime-pru/Analisis-de-datos-espaciales/releases/tag/spatial-analysis).

2. Archive the release in Zenodo (or a similar service):
The release has been archived in Zenodo [here](https://zenodo.org/record/7967543).

3. Check that the Zenodo deposit has the correct metadata, including the title (which should match the paper title) and author list (ensuring the list is correct and excluding those who made only minor fixes); you may also add the authors' ORCID.
All the metadata has been checked.

4. List the Zenodo DOI of the archived version here.
The Zenodo DOI is 10.5281/zenodo.7967543.

Additionally, I have saved the paper with the correct name "paper.md" and preserved the Spanish version to address the mentioned issue.

Let me know if there is anything else I can assist you with.

Thanks for your time.
