+++ packages.bib
@@ -3,31 +3,31 @@
   author = {{R Core Team}},
   organization = {R Foundation for Statistical Computing},
   address = {Vienna, Austria},
-  year = {2022},
+  year = {2023},
   url = {https://www.R-project.org/},
 }
 
 @Manual{R-bookdown,
   title = {bookdown: Authoring Books and Technical Documents with R Markdown},
   author = {Yihui Xie},
-  year = {2022},
-  note = {R package version 0.30},
+  year = {2023},
+  note = {R package version 0.35},
   url = {https://CRAN.R-project.org/package=bookdown},
 }
 
 @Manual{R-knitr,
   title = {knitr: A General-Purpose Package for Dynamic Report Generation in R},
   author = {Yihui Xie},
-  year = {2022},
-  note = {R package version 1.40},
+  year = {2023},
+  note = {R package version 1.43},
   url = {https://yihui.org/knitr/},
 }
 
 @Manual{R-rmarkdown,
   title = {rmarkdown: Dynamic Documents for R},
-  author = {JJ Allaire and Yihui Xie and Jonathan McPherson and Javier Luraschi and Kevin Ushey and Aron Atkins and Hadley Wickham and Joe Cheng and Winston Chang and Richard Iannone},
-  year = {2022},
-  note = {R package version 2.18},
+  author = {JJ Allaire and Yihui Xie and Christophe Dervieux and Jonathan McPherson and Javier Luraschi and Kevin Ushey and Aron Atkins and Hadley Wickham and Joe Cheng and Winston Chang and Richard Iannone},
+  year = {2023},
+  note = {R package version 2.24},
   url = {https://CRAN.R-project.org/package=rmarkdown},
 }
 
@@ -37,7 +37,7 @@
   publisher = {Chapman and Hall/CRC},
   address = {Boca Raton, Florida},
   year = {2016},
-  note = {ISBN 978-1138700109},
+  isbn = {978-1138700109},
   url = {https://bookdown.org/yihui/bookdown},
 }
 
@@ -60,7 +60,6 @@
   publisher = {Chapman and Hall/CRC},
   year = {2014},
   note = {ISBN 978-1466561595},
-  url = {http://www.crcpress.com/product/isbn/9781466561595},
 }
 
 @Book{rmarkdown2018,
@@ -69,7 +68,7 @@
   publisher = {Chapman and Hall/CRC},
   address = {Boca Raton, Florida},
   year = {2018},
-  note = {ISBN 9781138359338},
+  isbn = {9781138359338},
   url = {https://bookdown.org/yihui/rmarkdown},
 }
 
@@ -79,7 +78,7 @@
   publisher = {Chapman and Hall/CRC},
   address = {Boca Raton, Florida},
   year = {2020},
-  note = {ISBN 9780367563837},
+  isbn = {9780367563837},
   url = {https://bookdown.org/yihui/rmarkdown-cookbook},
 }
 
