+++ 05-analisis-espacial-II.Rmd
@@ -26,13 +26,13 @@ u &=\lambda Wu+\epsilon  \\
 \end{aligned}
 $$
 
-Siguiendo a Elhorst [-@Elhorst2006: 10], se le denomina modelo general justamente por le hecho de que incorporara todos los tipos de efectos espaciales posibles. Veamos con cuidado la expresión anterior para distinguir cada uno de sus términos y ver con claridad los diferentes tipos de efectos espaciales.
+Siguiendo a Elhorst [-@Elhorst2006: 10], se le denomina modelo general justamente por el hecho de que incorporara todos los tipos de efectos espaciales posibles. Veamos con cuidado la expresión anterior para distinguir cada uno de sus términos y ver con claridad los diferentes tipos de efectos espaciales.
 
 El término identificado como $Y$ no es otro que el vector que contiene la variable dependiente de nuestro modelo, por ejemplo, el número de casos positivos por COVID19; $X$ corresponde a la matriz de variables independientes que pueden ser, por ejemplo, el acceso a los servicios de salud y la proporción promedio de ocupantes por vivienda; el término $\beta$ corresponde al conjunto de parámetros a ser estimados, además, seguro que te resulta familiar el término $W$ que no es otra cosa que la estructura espacial definida a partir de la matriz de pesos espaciales; finalmente, listemos los diferentes efectos espaciales que aparecen en la expresión:
 
-i) efectos de interacción endógenos, es decir, cuando la variable dependiente reaparece como independiente pero a través de su rezago espacial ($WY$), aquí el sentido y magnitud de la interacción está dado por el parámetro $\rho$, llamado coeficiente autoregresivo espacial;  
-ii) efectos de interacción exógenos, es decir, cuando las variables independientes son incluidas en forma de sus rezagos espaciales ($WX$) en la que la relación es captada por $\theta$, también un parámetro a estimar;  
-iii) efectos de interacción dados por los términos de error de la regresión ($Wu$) e interesa el sentido y magnitud de $\lambda$, el coeficiente a estimar, que es llamado coeficiente de autocorrelación espacial.  
+i) Efectos de interacción endógenos, es decir, cuando la variable dependiente reaparece como independiente pero a través de su rezago espacial ($WY$), aquí el sentido y magnitud de la interacción está dado por el parámetro $\rho$, llamado coeficiente autoregresivo espacial;  
+ii) Efectos de interacción exógenos, es decir, cuando las variables independientes son incluidas en forma de sus rezagos espaciales ($WX$) en la que la relación es captada por $\theta$, también un parámetro o parámetros a estimar;  
+iii) Efectos de interacción dados por los términos de error de la regresión ($Wu$) y de ellos interesa el sentido y magnitud de $\lambda$, el coeficiente a estimar, que es llamado coeficiente de autocorrelación espacial.  
 
 A partir del modelo general es posible derivar otros, más simples, si se suponen determinados valores para los coeficientes $\rho$, $\theta$ y $\lambda$. Toda la constelación de modelos econométricos espaciales aparece en la figura 5.1 [@Elhorst2006,p. 7]. Por ejemplo, si partiendo del modelo general se asume que $\lambda=0$, se obtiene el denominado modelo espacial de Durbin; en tanto si $\rho=\theta=\lambda=0$, llegamos al conocido modelo clásico de regresión lineal (MCO). Observa con cuidado cada una de las diferentes alternativas de modelación econométrica espacial y trata de reflexionar sobre sus diferencias.
 
@@ -42,7 +42,7 @@ knitr::include_graphics("recursos 5/figura 1.jpg")
 
[…]

[…]
