### 2024-07-31 #173 @labarba
Hi @jaime-pru — Thanks for the ping. Allison has retired from the editorial board, as she has left academia for an industry position. Let me have a look at where this submission stands and get back to you...
