### 2024-07-31 #173 @labarba
The metadata of the Zenodo deposit has the object classified as "[Computer software]"—are you sure that is what you want?

Also, please double check all author names and ORCID and confirm that they are correct. Check also the references, an add the DOI to the `rgeoda` package if correct.
