### 2024-07-31 #173 @labarba
Please check the following in the paper:

- a reference to `@tidyverse` is not rendering properly (p.1, par.3)
- "a very synthetic review" should be edited, as "synthetic" in English does not mean short or abridged
- (p.2, par.5) "within the economy. training" needs an edit (should it be a comma?)
- (p.3) "GitHub repository, here" — could you replace the hyperlinked word "here" but the full inline URL? (you can make it an active hyperlink if you wish)
- reference with author "IMCO…" —please check that it is rendered properly; e.g., it shows "social, I." as co-author. I suspect you need to use brackets `{}` to protect the author field.
