+++ paper.md
@@ -30,7 +30,7 @@ Spatial Analysis with R: wrangling, visualization and econometric models (Análi
 
 The objective of the book is to guide the student, from a basic practical approach, in the knowledge and handling of spatial information exploration, analysis and modeling techniques through the use of R [@RSoftware] and RStudio, which are, respectively, a program computer science and programming language focused on statistical analysis and information visualization, on the one hand, and an integrated development environment (IDE), on the other.  
 
-The course is presented in the form of an electronic book and it is structured in five chapters of gradual learning. In Chapter 1, the basics of R and RStudio are introduced using the exploratory data analysis approach [@HSM2014] not only with the basic R package, but also with the popular set of tools provided by tidyverse [@ tidyverse], that is, the student is completely introduced to the use of the software to propose and solve questions related to the information structure. Chapter 2 shows how to create various types of choropleth maps and the enormous styling customization flexibility for this purpose via the tmap [@tmap] package. Chapter 3 presents how to carry out an exploratory analysis of spatial data where the student will find how to define the interrelationships that take place in space through the construction of spatial weight matrices and she will learn about spatial autocorrelation and its implications in the information analysis, through various packages such as spdep [@spdep] and rgeoda [@rgeoda]. Meanwhile, in Chapter 4, a very synthetic review of simple regression models is presented, emphasizing the autocorrelation problem that could occur when estimating a linear model with spatial data. Finally, in Chapter 5, two of the different spatial econometric modeling alternatives available in R it is shown, with spatialreg [@spatialreg].  
+The course is presented in the form of an electronic book and it is structured in five chapters of gradual learning. In Chapter 1, the basics of R and RStudio are introduced using the exploratory data analysis approach [@HSM2014] not only with the basic R package, but also with the popular set of tools provided by tidyverse [@tidyverse], that is, the student is completely introduced to the use of the software to propose and solve questions related to the information structure. Chapter 2 shows how to create various types of choropleth maps and the enormous styling customization flexibility for this purpose via the tmap [@tmap] package. Chapter 3 presents how to carry out an exploratory analysis of spatial data where the student will find how to define the interrelationships that take place in space through the construction of spatial weight matrices and she will learn about spatial autocorrelation and its implications in the information analysis, through various packages such as spdep [@spdep] and rgeoda [@rgeoda]. Meanwhile, in Chapter 4, a very brief review of simple regression models is presented, emphasizing the autocorrelation problem that could occur when estimating a linear model with spatial data. Finally, in Chapter 5, two of the different spatial econometric modeling alternatives available in R it is shown, with spatialreg [@spatialreg].  
 
 The logic of each chapter integrates three elements: i) explanation of the fundamental concepts covered, ii) the use of real information in the software that serves to illustrate the highlighted concepts, iii) exercises proposed for the student to delve into the topics exposed.  
 
@@ -46,7 +46,7 @@ Since January 2022, when the integrated version of this series of materials was 
 
 # Statement of Need
 
-Regional sciences are characterized by their multidisciplinary character and their solid quantitative support [@Fischer2014]. Although it is true that in all economics Majors we find a solid repertoire of quantitative instruments: mathematics, statistics and, of course, econometrics, the necessary tools for the analysis of reality from a spatial and regional perspective are still scarce within the economy. training at the undergraduate level. [Analysis of spatial data with R](https://jaime-pru.github.io/Analisis-de-datos-espaciales/index.html) intends to be a contribution, albeit minimal, to remedy this situation.  
[…]
