+++ paper_es.md
@@ -25,21 +25,21 @@ bibliography: paper.bib
 ---
 #Resumen
 
-Análisis Espacial con R: manejo, visualización y modelos econométricos es un [curso abierto completo](https://github.com/jaime-pru/Analisis-de-datos -espaciales) en español para el análisis de información espacial en software R. El curso está dirigido a estudiantes de pregrado de economía de habla hispana interesados en desarrollar habilidades técnicas para el análisis cuantitativo que requieren las ciencias regionales [@Fischer2014] y el enfoque espacial de la economía [@Fujita2000], pero podría ser útil para otras estudiantes de ciencias sociales atraídas por el tema.
+Análisis Espacial con R: manejo, visualización y modelos econométricos es un [curso abierto completo](https://github.com/jaime-pru/Analisis-de-datos -espaciales) en español para el análisis de información espacial en software R. El curso está dirigido a estudiantes de pregrado de economía de habla hispana interesados en desarrollar habilidades técnicas para el análisis cuantitativo que requieren las ciencias regionales [@Fischer2014] y el enfoque espacial de la economía [@Fujita2000], pero podría ser útil para otros estudiantes de ciencias sociales atraídos por el tema.
 
-El objetivo del libro es guiar a la estudiante, desde un enfoque fundamentalmente práctico, en el conocimiento y manejo de técnicas de exploración, análisis y modelado de información espacial mediante el uso de R [@RSoftware] y RStudio que son, respectivamente, un programa informático y lenguaje de programación enfocado al análisis estadístico y visualización de información, por un lado, y un entorno de desarrollo integrado (IDE), por otro.
+El objetivo del libro es guiar al estudiante, desde un enfoque fundamentalmente práctico, en el conocimiento y manejo de técnicas de exploración, análisis y modelado de información espacial mediante el uso de R [@RSoftware] y RStudio que son, respectivamente, un programa informático y lenguaje de programación enfocado al análisis estadístico y visualización de información, por un lado, y un entorno de desarrollo integrado (IDE), por otro.
 
-El curso se presenta en forma de libro electrónico y está estructurado en cinco capítulos de aprendizaje gradual. En el Capítulo 1, se presentan los elementos básicos de R y RStudio mediante el uso del enfoque de análisis exploratorio de datos [@HSM2014] no sólo con el paquete básico de R, sino también con el popular conjunto de herramientas que provee tidyverse [@tidyverse], es decir, se introduce completamente al estudiante en el uso del software para proponer y resolver preguntas relacionadas con la estructura información. El Capítulo 2 muestra cómo crear varios tipos de mapas de coropletas y la enorme flexibilidad de personalización de estilo que tiene R para este propósito a través del paquete tmap [@tmap]. El Capítulo 3 presenta la forma de realizar un análisis exploratorio de datos espaciales donde el estudiante encontrará cómo definir las interrelaciones que se dan en el espacio a través de la construcción de matrices de pesos espaciales y aprenderá sobre la autocorrelación espacial y sus implicaciones en el análisis de la información, todo ello a través de diversos paquetes como spdep [@spdep] y rgeoda [@rgeoda]. Mientras tanto, en el Capítulo 4, se presenta una revisión muy sintética de los modelos de regresión simple, enfatizando el problema de autocorrelación que podría ocurrir al estimar un modelo lineal con datos espaciales. Finalmente, en el Capítulo 5, se muestran dos de las diferentes alternativas de modelado econométrico espacial disponibles en R, con spatialreg [@spatialreg].
[…]
