+++ paper.bib
@@ -8,10 +8,10 @@
 }
 
 @report{IMCO2015,
-   author = {IMCO and COMCE and Inglés para la competitividad y la movilidad social},
+   author = {},
    city = {Ciudad de México},
-   institution = {Instituto Mexicano para la Competitividad, A.C.},
-   title = {Inglés es posible Propuesta de una Agenda Nacional},
+   institution = {Instituto Mexicano para la Competitividad, A.C.; Inglés para la competitividad y la movilidad social; Consejo Empresarial Mexicano de Comercio Exterior, Inversión y Tecnología},
+   title = {Inglés es posible. Propuesta de una Agenda Nacional},
    url = {https://imco.org.mx/wp-content/uploads/2015/04/2015_Documento_completo_Ingles_es_posible.pdf},
    year = {2015},
 }
