+++ paper.bib
@@ -26,11 +26,13 @@
 
 @book{HSM2014,
    author = {C Croarkin and P Tobias},
-   journal = {Retrieved January},
-   title = {NIST/SEMATECH e-handbook of statistical methods},
-   volume = {1},
+   title = {Engineering statistics handbook: e-handbook of statistical methods},
+   publisher = {NIST/SEMATECH},
    year = {2014},
+   doi = {10.18434/M32189},
 }
+
+
 
 @Article{tidyverse,
   title = {Welcome to the {tidyverse}},
@@ -72,6 +74,7 @@
   year = {2022},
   note = {R package version 0.0.9},
   url = {https://CRAN.R-project.org/package=rgeoda},
+  doi = {10.32614/CRAN.package.rgeoda},
   }
 
 @book{Montgomery2012,
