### 2024-08-08 #173 @jaime-pru
Hi, again, @labarba. I think the last details have been resolved and we could proceed.
