### 2024-10-10 #173 @labarba
@openjournals/dev Could you help us out here? (Some sort of API error: `422 - Invalid request.` and `"sha" wasn't supplied.`)
