### 2024-10-10 #173 @arfon
@jaime-pru – could you please merge this PR? https://github.com/jaime-pru/Analisis-de-datos-espaciales/pull/16
