### 2024-10-10 #173 @arfon
@labarba – I *think* https://github.com/jaime-pru/Analisis-de-datos-espaciales/pull/16 should fix this (you should be able to do `@editorialbot reaccept` once merged)
