### 2024-10-10 #173 @labarba
@arfon : by doing a `reaccept` we didn't get the acceptance actions completed here, e.g., the link to landing page, the tweet, the citation recs, etc. Should we run `accept` again?
