### 2024-10-10 #173 @labarba
@openjournals/dev We're still stumped on how to complete all acceptance steps on this submission. Help?
