### 2024-10-10 #173 @arfon
> @arfon : by doing a reaccept we didn't get the acceptance actions completed here, e.g., the link to landing page, the tweet, the citation recs, etc. Should we run accept again?

Unfortunately that's a limitation of the situation we arrived at here. The paper is published, but I can't get these to generate now sorry.
