### 2024-10-11 #173 @labarba
Hi @jaime-pru — Your JOSE paper is published here: https://doi.org/10.21105/jose.00173
