### 2024-10-11 #173 @jaime-pru
@labarba, I want to thank you for leading this process to completion. I’m also very grateful to @allisonhorst for her time.

A special thanks to @nmorandeira and @hernandezcortes for their valuable feedback. Their detailed comments and suggestions have greatly enriched the proposal.

¡Muchas gracias por todo!
